% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADaMDataPelicanInText}
\alias{ADaMDataPelicanInText}
\title{Example ADaM data for the 'inTextSummaryTable' package, currently 
extracted from the \code{ADaM} datasets for the
Pelican study ('study').}
\format{list with data.frame for each dataset}
\description{
Example ADaM data for the 'inTextSummaryTable' package, currently 
extracted from the \code{ADaM} datasets for the
Pelican study ('study').
}
\author{
Laure Cougnaud
}
\keyword{data}
