% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{checkVarLabInclude}
\alias{checkVarLabInclude}
\title{Check the \code{varLabInclude} variable.}
\usage{
checkVarLabInclude(var, varLabInclude = length(var) > 1)
}
\arguments{
\item{var}{String with variable to check.}

\item{varLabInclude}{Logical, if TRUE
the name of the summary statistic variable(s) (\code{var})
are included in the table.
This is automatically set to TRUE if more than one variable(s) 
and is specified, and FALSE if only one variable is specified.}
}
\value{
(Updated) \code{varLabInclude}
}
\description{
This function ensures that:
\itemize{
\item{variable name is included if more than one variable
are specified
}
\item{variable name is not included if no variable is specified}
}
}
\author{
Laure Cougnaud
}
