% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTable-combine.R
\name{combine}
\alias{combine}
\title{Combine objects}
\usage{
combine(...)
}
\arguments{
\item{...}{Extra parameters for the corresponding
method.}
}
\description{
Combine objects
}
\seealso{
\code{\link{combine.summaryTable}}
to combine \code{\link{summaryTable}} objects.
}
