% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTable-combine.R
\name{combineColTotal}
\alias{combineColTotal}
\title{Combine the column total for a combined summary table}
\usage{
combineColTotal(summaryTable, attrs)
}
\arguments{
\item{summaryTable}{Combined summary table}

\item{attrs}{Nested list with attributes of each
summary table.}
}
\value{
summary table, with combined total columns
}
\description{
Combine the column total for a combined summary table
}
\author{
Laure Cougnaud
}
