% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSummaryStatistics.R
\name{computeSummaryStatistics}
\alias{computeSummaryStatistics}
\title{Compute summary statistics for a specific dataset and variables of interest}
\usage{
computeSummaryStatistics(data, var = "AVAL", colVar = NULL,
  rowVar = NULL, subjectVar = "USUBJID")
}
\arguments{
\item{data}{data.frame with data}

\item{var}{string, variable of \code{data} with variable to use,
'AVAL' by default}

\item{colVar}{variable(s) of \code{data} used 
for grouping in column in the final table. The total 
for each subgroup across \code{rowVar} is computed.}

\item{rowVar}{variable(s) of \code{data} used for
grouping in row in the final table.}

\item{subjectVar}{string, variable of \code{data} with subject ID,
'USUBJID' by default}
}
\value{
data.frame with summary statistics in columns:
\itemize{
\item{'N': }{number of subjects}
\item{'Mean': }{mean of \code{var}}
\item{'SD': }{standard deviation of \code{var}}
\item{'SE': }{standard error of \code{var}}
\item{'Median': }{median of \code{var}}
\item{'Min': }{minimum of \code{var}}
\item{'Max': }{maximum of \code{var}}
}
}
\description{
Compute summary statistics for a specific dataset and variables of interest
}
\author{
Laure Cougnaud
}
