% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSummaryStatisticsTable.R
\name{computeSummaryStatisticsByRowColVar}
\alias{computeSummaryStatisticsByRowColVar}
\title{Compute summary statistics by specified \code{rowVar} and \code{colVar}}
\usage{
computeSummaryStatisticsByRowColVar(data, var = NULL,
  varLab = getLabelVar(var = var, data = data, labelVars = labelVars),
  varInclude0 = FALSE, varTotalInclude = FALSE, type = "auto",
  rowVar = NULL, rowInclude0 = FALSE, rowVarDataLevels = NULL,
  colVar = NULL, colInclude0 = FALSE, colVarDataLevels = NULL,
  subjectVar = "USUBJID", labelVars = NULL, statsExtra = NULL)
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{String with variable of \code{data} to sort.}

\item{varLab}{Named character vector with label for each variable 
specified in \code{var}.
By default, extracted from the \code{labelVars}.}

\item{varInclude0}{Logical, if TRUE (FALSE by default)
include rows with no counts for the count \code{var} or \code{varFL} variable(s).}

\item{varTotalInclude}{Should the total across all categories of \code{var} 
be included for the count table?
Either:
\itemize{
\item{logical of length 1, if TRUE (FALSE by default) include the total for all \code{var}}
\item{a character vector containing \code{var} for which the total should be included}
}}

\item{rowVar}{Variable(s) of \code{data} used for
grouping in row in the final table.}

\item{rowInclude0}{Logical, if TRUE (FALSE by default),
include rows with no records, based on all combinations 
of the \code{rowVar} (assuming nested variable(s)).}

\item{rowVarDataLevels}{Data.frame with levels to consider for the \code{rowVar} variable(s)
(in case of non nested row variable(s)).}

\item{colVar}{Character vector with variable(s) used for the columns.
If multiple variables are specified, the variables should be sorted in hierarchical order,
and are included in multi-columns layout.}

\item{colInclude0}{Logical, if TRUE (FALSE by default),
include columns with no records, based on all combinations 
of the \code{columnVar} (assuming nested variable(s)).
If variable(s) are not nested, possible combinations
can be specified via \code{colVarDataLevels}.}

\item{colVarDataLevels}{Data.frame with levels to consider for the \code{colVar} variable(s)
(in case of non nested column variable(s)).}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
data.frame of class 'countTable' or 'summaryTable',
depending on the 'type' parameter; with statistics in columns,
either if \code{type} is:
\itemize{
\item{'summaryTable': }{
\itemize{
\item{'N': }{number of subjects}
\item{'Mean': }{mean of \code{var}}
\item{'SD': }{standard deviation of \code{var}}
\item{'SE': }{standard error of \code{var}}
\item{'Median': }{median of \code{var}}
\item{'Min': }{minimum of \code{var}}
\item{'Max': }{maximum of \code{var}}
\item{'m': }{number of records}
}
}
\item{'countTable': }{
\itemize{
\item{'N': }{number of subjects}
\item{'m': }{number of records}
}
}
}
}
\description{
Compute summary statistics by specified \code{rowVar} and \code{colVar}
}
\author{
Laure Cougnaud
}
\keyword{internal}
