% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSummaryStatistics.R
\name{computeSummaryStatisticsOneVar}
\alias{computeSummaryStatisticsOneVar}
\title{Get summary statistics of interest of an unique variable of interest.}
\usage{
computeSummaryStatisticsOneVar(data, var = NULL,
  subjectVar = "USUBJID", type = "summaryTable",
  filterEmptyVar = TRUE)
}
\arguments{
\item{data}{data.frame with data}

\item{var}{string, variable of \code{data} with variable to compute statistics on,
only used (and required) if \code{type} is 'summaryTable'.
Missing values, if present, are filtered.}

\item{subjectVar}{string, variable of \code{data} with subject ID,
'USUBJID' by default}

\item{type}{string with type of summary table: 'summaryTable' 
(by default) or 'countTable'}

\item{filterEmptyVar}{logical (TRUE by default), should the summary statistics be filtered
in case \code{var} is empty.}
}
\value{
data.frame with summary statistics in columns,
depending if \code{type} is:
\itemize{
\item{'summary': }{
\itemize{
\item{'N': }{number of subjects }
\item{'m': }{number of records}
\item{'Mean': }{mean of \code{var}}
\item{'SD': }{standard deviation of \code{var}}
\item{'SE': }{standard error of \code{var}}
\item{'Median': }{median of \code{var}}
\item{'Min': }{minimum of \code{var}}
\item{'Max': }{maximum of \code{var}}
}
}
\item{'count': }{
\itemize{
\item{'N': }{number of subjects}
\item{'m': }{number of records}
}
}
}
}
\description{
Get summary statistics of interest of an unique variable of interest.
}
\author{
Laure Cougnaud
}
