% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{convertRowColVarToFactor}
\alias{convertRowColVarToFactor}
\title{Convert \code{rowVar} and \code{colVar} in \code{data} to factor}
\usage{
convertRowColVarToFactor(data, rowVar = NULL, colVar = NULL)
}
\arguments{
\item{data}{Data.frame with data.}

\item{rowVar}{Character vector with variable(s) used for the rows.}

\item{colVar}{Character vector with variable(s) used for the columns.}
}
\value{
Updated \code{data}
}
\description{
Convert \code{rowVar} and \code{colVar} in \code{data} to factor
}
\author{
Laure Cougnaud
}
