% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT-convertSummaryStatisticsTableToDT.R
\name{convertSummaryStatisticsTableToDT}
\alias{convertSummaryStatisticsTableToDT}
\title{Convert summary table to DT}
\usage{
convertSummaryStatisticsTableToDT(summaryTable, rowVar,
  statsLayout = c("row", "col", "rowInSepCol"),
  statsValueLab = "StatisticValue", rowVarInSepCol = NULL,
  title = NULL, statsVar = NULL, expandVar = NULL)
}
\arguments{
\item{rowVar}{Character vector with variable(s) used for the rows.
If multiple variables are specified, the variables should be sorted in hierarchical order.
The variables are included in rows, excepted if specified in \code{rowVarInSepCol}.}

\item{statsLayout}{String with layout for the statistics names 
(in case more than one statistic is included), among:
\itemize{
\item{row: }{Statistics are included in rows in the first column of the table}
\item{'col': }{Statistics are included in columns (last row of the header).
This option is not compatible with categorical variable(s).}
\item{'rowInSepCol': }{Statistics are included in rows, but in a separated column than
the \code{rowVar} variable(s)}
}}

\item{rowVarInSepCol}{Variable(s) of \code{rowVar} which should be 
included in separated column in the table, NULL by default. 
To include the groups within a \code{var} variable in a separated column, set: rowVarInSepCol == 'variableGroup'.
This is only available if \code{rowVar} if not specified.
Note that the row total (if \code{rowTotalInclude} is TRUE) is computed 
separately by this variable.}

\item{statsVar}{Character vector with columns of \code{summaryTable} with
statistic variables.}

\item{expandVar}{Character vector with variables of the summary table which
should be expanded in the data (only for 'DT' output).}
}
\value{
\code{\link[DT]{data.tables}}
}
\description{
Convert summary table to DT
}
\author{
Laure Cougnaud
}
