% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{convertVarFlag}
\alias{convertVarFlag}
\title{Convert flag variable to a format such as only the flagged records are counted
in the summary table.}
\usage{
convertVarFlag(x)
}
\arguments{
\item{x}{Character or factor variable with flag variable,
should contain elements: 'Y' and 'N', or '' (for missing value).}
}
\value{
Formatted factor variable with levels: 'Y' & 'N'.
Empty strings have been converted to NA.
}
\description{
Convert flag variable to a format such as only the flagged records are counted
in the summary table.
}
\author{
Laure Cougnaud
}
