% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{convertVarRowVarColVarToFactor}
\alias{convertVarRowVarColVarToFactor}
\title{Convert \code{rowVar}, \code{colVar} and character \code{var} in \code{data} to factor}
\usage{
convertVarRowVarColVarToFactor(data, rowVar = NULL, colVar = NULL, var = NULL)
}
\arguments{
\item{data}{Data.frame with data.}

\item{rowVar}{Character vector with variable(s) used for the rows.}

\item{colVar}{Character vector with variable(s) used for the columns.}

\item{var}{Character vector with variable(s) used for the summary statistics.}
}
\value{
Updated \code{data}
}
\description{
Convert \code{rowVar}, \code{colVar} and character \code{var} in \code{data} to factor
}
\author{
Laure Cougnaud
}
