% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSummaryStatisticsTable.R
\name{convertVarToFactorWithOrder}
\alias{convertVarToFactorWithOrder}
\title{Convert a variable to a factor with levels in specified order.}
\usage{
convertVarToFactorWithOrder(data, var, otherVars = NULL,
  method = c("auto", "alphabetical", "total"), totalVar = "N")
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{String with variable of \code{data} to sort.}

\item{otherVars}{Character vector with other variable(s) of \code{data}
considered in the specific dimension.}

\item{totalVar}{String with variable of \code{data} considered in case \code{type} is 'total'.}

\item{type}{String with ordering method to use, either:
\itemize{
\item{'auto': }{if \code{var} is a factor, keep its order, otherwise alphabetically}
\item{'alphabetical': }{\code{var} is order in alphabetical order}
\item{'total': }{\code{var} is ordered based on the \code{totalVar} variable, in decreasing order.
The total count is extracted from the rows with all \code{otherVars} equal to 'Total'.
If none, consider all rows.
}
}}
}
\value{
Factor \code{var} variable of \code{data} with specified order.
}
\description{
Convert a variable to a factor with levels in specified order.
}
\author{
Laure Cougnaud
}
