% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSummaryStatisticsTableToFlextable.R
\name{convertVectToBinary}
\alias{convertVectToBinary}
\title{Convert vector to a bincode of 0/1
based on consecutive values in the vector.}
\usage{
convertVectToBinary(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
Integer vector of same length than \code{x}.
}
\description{
Convert vector to a bincode of 0/1
based on consecutive values in the vector.
}
\author{
Laure Cougnaud
}
