% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{cv}
\alias{cv}
\title{Compute coefficient of variation.}
\usage{
cv(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Logical, should NA value(s) be removed (FALSE by default)?}
}
\value{
Numeric vector of length 1 with coefficient of variation.
}
\description{
Compute coefficient of variation.
}
\author{
Laure Cougnaud
}
