% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSummaryStatisticsTable.R
\name{export}
\alias{export}
\title{Export an object}
\usage{
export(...)
}
\arguments{
\item{...}{Extra parameters for the corresponding
method.}
}
\description{
Export an object
}
\seealso{
\code{\link{export.summaryTable}}
to export \code{\link{summaryTable}} objects.
}
