% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSummaryStatisticsTableToFlextable.R
\name{exportFlextableToDocx}
\alias{exportFlextableToDocx}
\title{Export flextable to docx filr}
\usage{
exportFlextableToDocx(object, file, landscape = FALSE)
}
\arguments{
\item{object}{\code{\link[flextable]{flextable}} object, or list of such objects}

\item{file}{String with path of the file where the table should be exported.
If NULL, the summary table is not exported but only returned as output.}

\item{landscape}{Logical, if TRUE the file is in landscape format.}
}
\value{
no returned value, the \code{object} is exported to a docx file.
}
\description{
Export flextable to docx filr
}
\author{
Laure Cougnaud
}
