% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSummaryStatisticsTable.R
\name{exportSummaryStatisticsTable}
\alias{exportSummaryStatisticsTable}
\title{Export a summary table in \code{docx} format.}
\usage{
exportSummaryStatisticsTable(summaryTable,
  rowVar = getAttribute(summaryTable, "rowVar"),
  rowVarLab = getAttribute(summaryTable, "rowVarLab", default =
  getLabelVar(rowVar, labelVars = labelVars)),
  rowVarInSepCol = getAttribute(summaryTable, "rowVarInSepCol"),
  rowTotalInclude = getAttribute(summaryTable, "rowTotalInclude", default
  = FALSE), rowTotalLab = NULL,
  rowSubtotalInclude = getAttribute(summaryTable, "rowSubtotalInclude",
  FALSE), colVar = getAttribute(summaryTable, "colVar"),
  labelVars = NULL, file = NULL, landscape = (style ==
  "presentation"), margin = 1, rowPadBase = 4, title = NULL,
  footer = NULL, outputType = c("flextable", "data.frame"),
  statsLayout = c("row", "col", "rowInSepCol"), style = "report")
}
\arguments{
\item{summaryTable}{Summary table, created with the \code{\link{computeSummaryStatisticsTable}} function.}

\item{rowVar}{Character vector with variable(s) used for the rows.
If multiple variables are specified, the variables should be sorted in hierarchical order.
The variables are included in rows, excepted if specified in \code{rowVarInSepCol}.}

\item{rowVarLab}{Label for the \code{rowVar} variable(s).}

\item{rowVarInSepCol}{Variable(s) of \code{rowVar} which should be 
included in separated column in the table.
NULL by default. 
This is only available if \code{rowVar} if not NULL.}

\item{rowTotalInclude}{Logical, if TRUE (FALSE by default) include the total
counts across rows in a separated row.}

\item{rowTotalLab}{label for the row with total}

\item{rowSubtotalInclude}{Logical, if TRUE (FALSE by default) include also the total for each
group of the nested variable specified in \code{rowVar}}

\item{colVar}{Character vector with variable(s) used for the columns.
If multiple variables are specified, the variables should be sorted in hierarchical order,
and are included in multi-columns layout.}

\item{labelVars}{Named string with variable labels (names are the variable code).}

\item{file}{String with path of the file where the table should be exported.
If NULL, the summary table is not exported but only returned as output.}

\item{landscape}{Logical, if TRUE (by defaut) the table is presented in landscape
format.}

\item{margin}{Margin in the document in inches.}

\item{rowPadBase}{Base padding for row (number of spaces)}

\item{title}{Character vector with title(s) for the table.
Set to NULL (by default) if no title should be included.}

\item{footer}{Character vector with footer(s) for the table.
Set to NULL (by default) of no footer should be included.}

\item{outputType}{String with output type, 'data.frame' or 'flextable'.}

\item{statsLayout}{String with layout for the statistics names 
(in case more than one statistic is included), among:
\itemize{
\item{row: }{Statistics are included in rows in the first column of the table}
\item{'col': }{Statistics are included in columns (last row of the header)}
\item{'rowInSepCol': }{Statistics are included in rows, but in a separated column than
the \code{rowVar} variable(s)}
}}

\item{style}{string with table style in case \code{outputType} is 'flextable',
either 'report' or 'presentation'}
}
\value{
Depending on the \code{outputType}:
\itemize{
\item{'flextable': }{\code{\link[flextable]{flextable}} object with summary table}
\item{'data.frame': }{data.frame with summary table}
}
If \code{summaryTable} is a list of summary tables,
returns a list of corresponding summary tables in long format.
}
\description{
Export a summary table in \code{docx} format.
}
\author{
Laure Cougnaud
}
