% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT-exportSummaryStatisticsTableToDT.R
\name{exportSummaryStatisticsTableToDT}
\alias{exportSummaryStatisticsTableToDT}
\title{Convert summary table to DT}
\usage{
exportSummaryStatisticsTableToDT(summaryTable,
  rowVar = getAttribute(summaryTable, "rowVar"),
  rowVarLab = getAttribute(summaryTable, "rowVarLab", default =
  getLabelVar(rowVar, labelVars = labelVars)),
  rowVarInSepCol = getAttribute(summaryTable, "rowVarInSepCol"),
  statsVar = getAttribute(summaryTable, "statsVar"),
  statsLayout = c("row", "col", "rowInSepCol"),
  statsValueLab = "StatisticValue", title = NULL, expandVar = NULL,
  noEscapeVar = NULL, barVar = NULL, pageDim = NULL,
  labelVars = NULL)
}
\arguments{
\item{summaryTable}{Summary table, created with the \code{\link{computeSummaryStatisticsTable}} function.}

\item{rowVar}{Character vector with variable(s) used for the rows.
If multiple variables are specified, the variables should be sorted in hierarchical order.
The variables are included in rows, excepted if specified in \code{rowVarInSepCol}.}

\item{rowVarLab}{Label for the \code{rowVar} variable(s).}

\item{rowVarInSepCol}{Variable(s) of \code{rowVar} which should be 
included in separated column in the table, NULL by default. 
To include the groups within a \code{var} variable in a separated column, set: rowVarInSepCol == 'variableGroup'.
This is only available if \code{rowVar} if not specified.
Note that the row total (if \code{rowTotalInclude} is TRUE) is computed 
separately by this variable.}

\item{statsVar}{Character vector with columns of \code{summaryTable} with
statistic variables.}

\item{statsLayout}{String with layout for the statistics names 
(in case more than one statistic is included), among:
\itemize{
\item{row: }{Statistics are included in rows in the first column of the table}
\item{'col': }{Statistics are included in columns (last row of the header).
This option is not compatible with categorical variable(s).}
\item{'rowInSepCol': }{Statistics are included in rows, but in a separated column than
the \code{rowVar} variable(s)}
}}

\item{statsValueLab}{String with label for the statistic value, 
in case no \code{colVar} is specified: 'StatisticValue' by default.}

\item{title}{Character vector with title(s) for the table.
Set to NULL (by default) if no title should be included.
If vector > 1, specified for each element of \code{byVar} (in order of the levels).}

\item{expandVar}{Character vector with variables of the summary table which
should be expanded in the data (only for 'DT' output).}

\item{noEscapeVar}{Character vector with variables of \code{summaryTable}
which shouldn't be escaped in the table (e.g. containing URLs).}

\item{barVar}{Character vector with variables of \code{summaryTable}
that should be represented as a bar.}

\item{pageDim}{Numeric vector of length 2 with page width and height,
in number of rows (currently only
the height is used (e.g. \code{c(NA, 4)})}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
\code{\link[DT]{datatable}} object.
}
\description{
Convert summary table to DT
}
\author{
Laure Cougnaud
}
