% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable-formatSummaryStatisticsTableFt.R
\name{formatSummaryStatisticsTableFt}
\alias{formatSummaryStatisticsTableFt}
\title{Merge nested rows of a summary table
for a format compatible with \code{flextable}}
\usage{
formatSummaryStatisticsTableFt(data, rowVar, rowVarInSepCol,
  rowVarTotalInclude, statsLayout, statsVar, rowVarLab,
  rowVarTotalInSepRow = NULL, vline = c("none", "auto"),
  rowAutoMerge = TRUE, rowVarFormat = NULL)
}
\arguments{
\item{data}{data.frame with data}

\item{rowVar}{Character vector with variable(s) used for the rows.
If multiple variables are specified, the variables should be sorted in hierarchical order.
The variables are included in rows, excepted if specified in \code{rowVarInSepCol}.}

\item{rowVarInSepCol}{Variable(s) of \code{rowVar} which should be 
included in separated column in the table, NULL by default. 
To include the groups within a \code{var} variable in a separated column, set: rowVarInSepCol == 'variableGroup'.
This is only available if \code{rowVar} if not specified.
Note that the row total (if \code{rowTotalInclude} is TRUE) is computed 
separately by this variable.}

\item{rowVarTotalInclude}{Character vector with \code{rowVar}
for which to include the total for each group.}

\item{statsLayout}{String with layout for the statistics names 
(in case more than one statistic is included), among:
\itemize{
\item{row: }{Statistics are included in rows in the first column of the table}
\item{'col': }{Statistics are included in columns (last row of the header).
This option is not compatible with categorical variable(s).}
\item{'rowInSepCol': }{Statistics are included in rows, but in a separated column than
the \code{rowVar} variable(s)}
}}

\item{statsVar}{Character vector with columns of \code{summaryTable} with
statistic variables.}

\item{rowVarLab}{Label for the \code{rowVar} variable(s).}

\item{rowVarTotalInSepRow}{Character vector with \code{rowVarTotalInclude}
(not in \code{rowVarInSepCol}) for which the total should be included in a separated row labelled 'Total'.
Otherwise (by default) the total is included in the header row of each category.}
}
\value{
updated \code{data}
}
\description{
Merge nested rows of a summary table
for a format compatible with \code{flextable}
}
\author{
Laure Cougnaud
}
