% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSummaryStatisticsTableToFlextable.R
\name{formatSuperscriptToFlextable}
\alias{formatSuperscriptToFlextable}
\title{Format superscript in a flextable.
Superscript should be indicated as 'a^{b}' in the input summary table.}
\usage{
formatSuperscriptToFlextable(dataTable, ft, fontname = "Times",
  part = "body", fontsize = 8)
}
\arguments{
\item{dataTable}{data.frame with data used in table,
summary table for body or header data.frame for the header.}

\item{ft}{Corresponding \code{\link[flextable]{flextable}}.}

\item{fontname}{String with font name, 'Times' by default.}

\item{part}{string with part of the table to consider, 
see \code{\link[flextable]{display}}.}

\item{fontsize}{Integer with font size, 8 by default.}
}
\value{
\code{\link[flextable]{flextable}} with superscript.
}
\description{
Format superscript in a flextable.
Superscript should be indicated as 'a^{b}' in the input summary table.
}
\author{
Laure Cougnaud
}
