% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_stats.R
\name{geomCV}
\alias{geomCV}
\title{Compute geometric coefficient of variation.}
\usage{
geomCV(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Logical, should NA value(s) be removed (FALSE by default)?}
}
\value{
Numeric vector of length 1 with geometric coefficient of variation.
}
\description{
Compute geometric coefficient of variation.
}
\author{
Laure Cougnaud
}
