% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_stats.R
\name{geomMean}
\alias{geomMean}
\title{Compute geometric mean.}
\usage{
geomMean(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Logical, should NA value(s) be removed (FALSE by default)?}
}
\value{
Numeric vector of length 1 with geometric mean.
}
\description{
Compute geometric mean.
}
\author{
Laure Cougnaud
}
