% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{geomSD}
\alias{geomSD}
\title{Compute geometric standard deviation}
\usage{
geomSD(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Logical, should NA value(s) be removed (FALSE by default)?}
}
\value{
Numeric vector of length 1 with geometric mean.
}
\description{
Compute geometric standard deviation
}
\author{
Laure Cougnaud
}
