% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getAttribute}
\alias{getAttribute}
\title{Get specific attribute from a summaryTable or a list of summaryTables}
\usage{
getAttribute(summaryTable, name, default = NULL)
}
\arguments{
\item{name}{String with attribute name.}

\item{default}{Object with default value for attribute, 
NULL by default.}
}
\value{
Object with attribute, or \code{default} is doesn't
exists in \code{summaryTable}
}
\description{
Get specific attribute from a summaryTable or a list of summaryTables
}
\author{
Laure Cougnaud
}
\keyword{internal}
