% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSummaryStatisticsTableToFlextable.R
\name{getColorTable}
\alias{getColorTable}
\title{Get table color}
\usage{
getColorTable(style = c("report", "presentation"))
}
\arguments{
\item{style}{string with table style, either 'report' or 'presentation'}
}
\value{
Named character vector with color for the different parts of the table,
should at least contain font/background color for:
\itemize{
\item{header: }{'header'/'headerBackground'}
\item{body: }{'body' for body text and either 'bodyBackground' or 
'bodyBackground1'/'bodyBackground2': for body common background or background 
for alternate rows}
\item{footer: }{'footer'/'footerBackground'}
\item{line: }{'line'}
}
}
\description{
Get table color
}
\author{
Laure Cougnaud
}
