% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableColors.R
\name{getDefaultTableColors}
\alias{getDefaultTableColors}
\title{Get the default colors of the table}
\usage{
getDefaultTableColors(style = c("report", "presentation"))
}
\arguments{
\item{style}{style}
}
\description{
Get the default colors of the table
}
