% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSummaryStatisticsTableToFlextable.R
\name{getDimPage}
\alias{getDimPage}
\title{Return page dimension of interest}
\usage{
getDimPage(type = c("width", "height"), landscape = TRUE, margin = 1)
}
\arguments{
\item{type}{String dimension of interest, 'width' or 'height'.}

\item{landscape}{Logical, if TRUE (by defaut) the table is presented in landscape
format.}

\item{margin}{Margin in the document in inches.}
}
\value{
integer with dimension of interest
}
\description{
Return page dimension of interest
}
\author{
Laure Cougnaud
}
