% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getMaxNDecimals}
\alias{getMaxNDecimals}
\title{Get maximum number of decimals in a variable}
\usage{
getMaxNDecimals(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Integer with maximum number of decimals in a character vector.
}
\description{
Get maximum number of decimals in a variable
}
\author{
Laure Cougnaud
}
