% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimals.R
\name{getNDecimalsData}
\alias{getNDecimalsData}
\title{Get number of decimals based on the data in a numeric vector.
Note: NA is returned if the element is missing (NA).}
\usage{
getNDecimalsData(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector of same length than \code{x}
with the number of decimals.
}
\description{
Get number of decimals based on the data in a numeric vector.
Note: NA is returned if the element is missing (NA).
}
\author{
Laure Cougnaud
x <- c(12.345, -123.45, 1234.5, -12345)
getNDecimalsData(x)
}
