% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/flextable-convertSummaryStatisticsTableToFlextable.R
\name{getPatternPosition}
\alias{getPatternPosition}
\title{Get position(s) (start, end) of a pattern in a string.}
\usage{
getPatternPosition(x, startPattern, endPattern = "\\\\}", format)
}
\arguments{
\item{x}{String.}

\item{startPattern}{String with start pattern.}

\item{endPattern}{String with end pattern.}

\item{format}{String with type of formatting}
}
\value{
Matrix with columns: 'start' and 'end' with
start and end position(x) of the pattern, 'format' with
the \code{format} and 'pattern' with the regex pattern for the full match. 
NULL if no match.
}
\description{
Get position(s) (start, end) of a pattern in a string.
}
\author{
Laure Cougnaud
}
