% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_stats.R
\name{getStats}
\alias{getStats}
\title{Create custom statistic sets to be passed to the \code{stats} parameter
of the \code{\link{computeSummaryStatisticsTable}} function.}
\usage{
getStats(type = c("summary", "count", "n (\%)"))
}
\value{
(Optionally named) list of expression or call object of summary statistics of interest
}
\description{
Create custom statistic sets to be passed to the \code{stats} parameter
of the \code{\link{computeSummaryStatisticsTable}} function.
}
\author{
Laure Cougnaud
}
