% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSummaryStatistics.R
\name{getSummaryStatistics}
\alias{getSummaryStatistics}
\title{Get summary statistics of interest}
\usage{
getSummaryStatistics(data, var, subjectVar = "USUBJID",
  type = ifelse(is.numeric(data[, var]), "summaryTable", "countTable"))
}
\arguments{
\item{data}{data.frame with data}

\item{var}{string, variable of \code{data} with variable to use,
'AVAL' by default}

\item{subjectVar}{string, variable of \code{data} with subject ID,
'USUBJID' by default}

\item{type}{string with type of summary table: 'summaryTable' 
(by default if \code{var} is numeric) or 'countTable'}
}
\value{
data.frame with summary statistics in columns,
depending if \code{type} is:
\itemize{
\item{'summary': }{
\itemize{
\item{'N': }{number of subjects}
\item{'Mean': }{mean of \code{var}}
\item{'SD': }{standard deviation of \code{var}}
\item{'SE': }{standard error of \code{var}}
\item{'Median': }{median of \code{var}}
\item{'Min': }{minimum of \code{var}}
\item{'Max': }{maximum of \code{var}}
}
}
\item{'count': }{
\itemize{
\item{'N': }{number of subjects}
}
}
}
}
\description{
Get summary statistics of interest
}
\author{
Laure Cougnaud
}
