% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryStatisticsTable.R
\name{getSummaryStatisticsTable}
\alias{getSummaryStatisticsTable}
\title{Get summary statistics table}
\usage{
getSummaryStatisticsTable(data, var = "AVAL", rowVar = NULL,
  rowVarLab = getLabelVar(rowVar, labelVars = labelVars),
  colVar = NULL, subjectVar = "USUBJID", labelVars = NULL,
  landscape = FALSE, title = "Table: Descriptive statistics",
  file = "summaryStatisticsTable.docx")
}
\arguments{
\item{data}{data.frame with summary statistics to represent in the plot,
e.g. produced with the \code{\link{computeSummaryStatistics}}}

\item{var}{string, variable of \code{data} with variable to use,
'AVAL' by default}

\item{rowVar}{character vector with variable(s) of \code{data}
used for grouping in rows.}

\item{rowVarLab}{label for each variable of \code{rowVar}.}

\item{colVar}{string with variable of \code{data} used for grouping in column.}

\item{subjectVar}{string, variable of \code{data} with subject ID,
'USUBJID' by default}

\item{labelVars}{named string with variable labels (names are the variable code)}

\item{landscape}{logical, if TRUE (by defaut) the table is presented in landscape
format}

\item{title}{string with title for the table.
Set to NULL if no title should be included.}

\item{file}{string with path of the file where the table should be exported}
}
\description{
Get summary statistics table
}
\author{
Laure Cougnaud
}
