% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{hlineBottom}
\alias{hlineBottom}
\title{Include hline at the bottom of a row
(fix in case cells are merged)}
\usage{
hlineBottom(ft, data, iBase = 0, part = "body", bd = fp_border())
}
\arguments{
\item{ft}{flextable}

\item{data}{Data.frame with data on which the table (indices computation) is based on.}

\item{iBase}{Base index, 0 by default}
}
\value{
update flextable object
}
\description{
Include hline at the bottom of a row
(fix in case cells are merged)
}
\author{
Laure Cougnaud
}
