% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{interactionCustom}
\alias{interactionCustom}
\title{Compute the interaction between variable(s),
without propagating the missing values (if present in one of the variable)
unlike the behaviour with \code{\link{interaction}}.
This also ensure that the levels of the final interaction variables
are ordered similarly as the levels of the input \code{var} (if present).}
\usage{
interactionCustom(data, var)
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{Character vector with variable(s) to consider.}
}
\value{
factor with interaction between the input \code{var}
}
\description{
Compute the interaction between variable(s),
without propagating the missing values (if present in one of the variable)
unlike the behaviour with \code{\link{interaction}}.
This also ensure that the levels of the final interaction variables
are ordered similarly as the levels of the input \code{var} (if present).
}
\author{
Laure Cougnaud
}
