% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inTextSummaryTable-common-args-doc.R
\name{inTextSummaryTable-DT-args}
\alias{inTextSummaryTable-DT-args}
\title{Common arguments for the functionalities of the inTextSummaryTable package
for DT export.}
\arguments{
\item{expandVar}{(DT output) Character vector with variables of the summary table which
should be expanded in the data.}

\item{pageDim}{(DT output) Numeric vector of length 2 with page width and height,
in number of rows (currently only
the height is used (e.g. \code{c(NA, 4)})}

\item{noEscapeVar}{(DT output) Character vector with variables of \code{summaryTable}
which shouldn't be escaped in the table (e.g. containing URLs).}

\item{barVar}{(DT output) Character vector with variables of \code{summaryTable}
that should be represented as a bar.}

\item{file}{String with path of the file where the table should be exported.
The file should have the extension: '.docx'.
If NULL, the summary table is not exported but only returned as output.
If \code{byVar} is specified, each table is exported to a separated
file with the suffix: 'file_[i].docx' with i the index of the file.}
}
\value{
No return value, used for the documentation of 
R functions for 'DT' output
}
\description{
Common arguments for the functionalities of the inTextSummaryTable package
for DT export.
}
