% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inTextSummaryTable.R
\docType{package}
\name{inTextSummaryTable}
\alias{inTextSummaryTable}
\title{inTextSummaryTable: creation of in-text summary table}
\description{
Tables of summary statistics or count tables are created. 
These tables can be exported as in-text table to a Clinical Study Report 
(Word format), a topline presentation (PowerPoint format),
 or as interactive table to an html document.
}
\details{
\itemize{
\item{To get started with the package, see:\cr}{
\code{vignette("inTextSummaryTable-introduction", package = "inTextSummaryTable")}}
\item{To get example code for standard in-text tables created with the package, 
see:\cr}{
\code{vignette("inTextSummaryTable-standardTables", package = "inTextSummaryTable")}}
\item{The main function: \code{\link{getSummaryStatisticsTable}} enables
to create ready in-text table.}
}
}
