% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable-getListing.R
\name{mergeHeaderCols}
\alias{mergeHeaderCols}
\title{Merge columns and rows in the header of a flextable object.}
\usage{
mergeHeaderCols(x, mapping)
}
\arguments{
\item{x}{a \code{flextable} object}

\item{mapping}{a \code{data.frame} specyfing for each colname
content of the column.}
}
\value{
Updated flextable object with merged cells in the header.
}
\description{
This:
\itemize{
\item{(1) for each row: checks that subsequent columns are identical (and corresponding
sub-header overlap), so should be merged}
\item{(2) for each column: checks that subsequent rows are identical, so should be
merged}
\item{filters cells planned to be merged across rows (2) if they are already
merged across columns (1)}
}
}
\keyword{internal}
