% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{data}
\name{pageDimPresentation}
\alias{pageDimPresentation}
\title{Page dimension for powerpoints}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
pageDimPresentation
}
\description{
Page dimension for powerpoints
}
\keyword{datasets}
