% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{postProcessVarFlag}
\alias{postProcessVarFlag}
\title{Post-process the summary statistics table with variable flag.}
\usage{
postProcessVarFlag(summaryTable, varFlag)
}
\arguments{
\item{summaryTable}{Summary table as created internally in
\code{\link{computeSummaryStatisticsTable}}.}

\item{varFlag}{Character vector, subset of \code{var} with variable(s) 
of type 'flag' (with 'Y', 'N' or '' for empty/non specified value).
Only the counts for records flagged (with 'Y') are retained.}
}
\value{
Summary table with
}
\description{
This function is for internal use within
the \code{\link{computeSummaryStatisticsTable}}
function.
}
\details{
This includes:
\itemize{
\item{converting the records from a flag variable
for the 'variableGroup' variable 
from 'Y' to \code{NA_character_}
}
\item{filter records from a flag variable 
with variableGroup set as 'N'}
}
}
\author{
Laure Cougnaud
}
