% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{roundCustom}
\alias{roundCustom}
\title{Custom round function, rounding to the closest digits
(instead of rounding to the even number in case of 0.5)}
\usage{
roundCustom(x, digits = 0)
}
\arguments{
\item{x}{numeric vector to round}

\item{digits}{number of digits to consider, 0 by default}
}
\value{
rounded vector
}
\description{
Custom round function, rounding to the closest digits
(instead of rounding to the even number in case of 0.5)
}
\examples{
# numbers are rounded to the closest even number in case of .5 
# with the round 'base' function
round(0.45, 1)
# 'roundCustom' round to the closest number
roundCustom(0.45, 1)
# rounding is the same for uneven number:
round(0.55, 1) == roundCustom(0.55, 1)
}
\author{
stackoverflow question 6461209
}
