% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{roundCustomText}
\alias{roundCustomText}
\title{Custom round function, rounding to the closest digits
(instead of rounding to the even number in case of 0.5)}
\usage{
roundCustomText(x, digits = 0, format)
}
\arguments{
\item{x}{Numeric vector to round.}

\item{digits}{Integer with number of digits to consider, 0 by default.}

\item{format}{For backward compatibility. This parameter will be deprecated in the next package release.}
}
\value{
A character vector with the rounded number.
See the \code{glpgUtilityFct::roundCustom} for the rounding customization.
}
\description{
Custom round function, rounding to the closest digits
(instead of rounding to the even number in case of 0.5)
}
\examples{
# number of digits higher than number of decimal
roundCustomText(x = c(0.345, 0.567, -0.98), digits = 2)
# number of digits lower than number of decimal
roundCustomText(x = c(0.345, 0.567, -0.98), digits = 0)
# by default, 'digits' is 0!
roundCustomText(x = c(0.345, 0.567, -0.98))
}
\author{
Laure Cougnaud and Michela Pasetto
}
