% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{se}
\alias{se}
\title{Compute standard error of the mean}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Logical, should NA value(s) be removed (FALSE by default)?}
}
\value{
Numeric vector with standard error of the mean
}
\description{
Compute standard error of the mean
}
\author{
Laure Cougnaud
}
