% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/flextable-formatSummaryStatisticsTableFlextable.R
\name{smoothPadding}
\alias{smoothPadding}
\title{Smooth padding, e.g. remove padding bigger than 1}
\usage{
smoothPadding(pad)
}
\arguments{
\item{pad}{Integer vector with padding.}
}
\value{
Integer vector with 'smooth' padding.
}
\description{
Smooth padding, e.g. remove padding bigger than 1
}
\author{
Laure Cougnaud
}
\keyword{internal}
