% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileSummaryPlot.R
\name{subjectProfileSummaryPlot}
\alias{subjectProfileSummaryPlot}
\title{Plot subject summary profile}
\usage{
subjectProfileSummaryPlot(data, xVar = NULL, xLab = getLabelVar(xVar,
  labelVars = labelVars), meanVar = "Mean", seVar = "SE",
  yLab = paste(meanVar, if (!is.null(seVar)) paste("+-", seVar)),
  facetVar = NULL, facetScale = "free_y", colorVar = NULL,
  colorLab = getLabelVar(colorVar, labelVars = labelVars),
  colorPalette = NULL, labelVars = NULL, useLinetype = TRUE,
  linetypePalette = NULL, useShape = TRUE, shapePalette = NULL,
  jitter = NULL, title = NULL, ylim = NULL, xlim = NULL)
}
\arguments{
\item{data}{Data.frame with summary statistics to represent in the plot,
e.g. produced with the \code{\link{computeSummaryStatisticsTable}}.}

\item{xVar}{String, variable of \code{data} with variable for the x-axis.}

\item{xLab}{String with label for the \code{xVar}.}

\item{meanVar}{String, variable of \code{data} with the mean variable.}

\item{seVar}{String, variable of \code{data} with the standard error.}

\item{yLab}{String with label for the y-axis.}

\item{facetVar}{String, variable of \code{data} for facetting.}

\item{facetScale}{String with type of scale used for facetting, 'free_y' by default
(fixed scale in the x-axis and free in the y-axis).}

\item{colorVar}{String, variable of \code{data} for coloring.}

\item{colorLab}{String, label for \code{colorVar}, used in the legend.}

\item{colorPalette}{(named) Vector with color palette.}

\item{labelVars}{Named string with variable labels (names are the variable code).}

\item{useLinetype}{Logical, if TRUE (FALSE by default) use also linetype
to differenciate the variable specified via \code{colorVar} in the mean line.}

\item{linetypePalette}{Vector with linetype(s), in case \code{useLinetype} is TRUE.}

\item{jitter}{Numeric with jitter for the x-axis, only used if \code{colorVar} specified.}

\item{title}{String with title for the plot.}
}
\value{
\code{\link[ggplot2]{ggplot}} object
}
\description{
Plot subject summary profile
}
\author{
Laure Cougnaud
}
