% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileSummaryPlot.R
\name{subjectProfileSummaryTable}
\alias{subjectProfileSummaryTable}
\title{Plot a table with \code{ggplot} of a variable of interest}
\usage{
subjectProfileSummaryTable(data, xVar, text, xLim = NULL,
  colorVar = NULL, colorPalette = NULL,
  colorLab = getLabelVar(colorVar, labelVars = labelVars), xLab = NULL,
  textSize = GeomText$default_aes$size, labelVars = NULL,
  showLegend = TRUE, yAxisLabs = FALSE, xAxisLabs = NULL)
}
\arguments{
\item{data}{Data.frame with data for the table.}

\item{xVar}{String, variable of \code{data} with variable for the x-axis.}

\item{xLim}{Vector of the length 2 with limits for the x-axis.}

\item{colorVar}{String, variable of \code{data} for coloring.}

\item{colorPalette}{(named) Vector with color palette.}

\item{colorLab}{String, label for \code{colorVar}, used in the legend.}

\item{xLab}{String with label for the x-axis.}

\item{textSize}{Size for the text.}

\item{labelVars}{Named string with variable labels (names are the variable code).}

\item{showLegend}{Logical, should the legend be displayed? TRUE by default.}

\item{yAxisLabs}{Logical, if TRUE include the labels in the y-axis.}

\item{xAxisLabs}{Vector with labels for the x-axis if \code{xVar}
is discrete or vector with limits if continuous.}

\item{label}{Character vector with colnames of \code{data}
or expression based on colnames of \code{data} to extract
the text label.}
}
\value{
\code{\link[ggplot2]{ggplot}} object
}
\description{
Plot a table with \code{ggplot} of a variable of interest
}
\author{
Laure Cougnaud
}
