% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableColors.R
\name{tableColorsPresentation}
\alias{tableColorsPresentation}
\title{Colors for tables in a presentation style}
\usage{
tableColorsPresentation
}
\description{
Default colors are
\itemize{
\item{header: white text on a blue background}
\item{body: black text on a grey background}
\item{footer: black text on a white background.}
}
}
