% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableColors.R
\docType{data}
\name{tableColorsReport}
\alias{tableColorsReport}
\title{Colors for tables in a report style}
\format{
An object of class \code{character} of length 7.
}
\usage{
tableColorsReport
}
\description{
Default colors are black text on a white background.
}
\keyword{datasets}
