% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryTable-combine.R
\name{uniqueVarWithOrder}
\alias{uniqueVarWithOrder}
\title{Get unique variables with meaningful order.}
\usage{
uniqueVarWithOrder(...)
}
\arguments{
\item{...}{Lists}
}
\value{
Vector with unique and ordered elements.
}
\description{
The following framework is followed:
\enumerate{
\item{get the unique elements in the vectors}
\item{for each of this element: get the average order across
the different vectors}
\item{put \code{variable}, if present as second to last element}
\item{put \code{variableGroup}, if present, as last element}
\item{order the unique elements based on the extracted order}
}
}
\author{
Laure Cougnaud
}
