% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{writeTable}
\alias{writeTable}
\title{Custom function to write table to a text file}
\usage{
writeTable(x, file, ...)
}
\arguments{
\item{x}{Data.frame to export to the table.}

\item{file}{String with text file to export to.}

\item{...}{Any parameters passed to the \code{\link[utils]{write.table}} function.}
}
\value{
No returned value, the object \code{x} is exported to the specified \code{file}.
}
\description{
This function is mainly a wrapper on \code{\link[utils]{write.table}},
with the specific options:
\itemize{
\item{no rownames}
\item{no quoting}
\item{tab separator}
}
}
\author{
Laure Cougnaud
}
