% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineVariables.R
\name{combineVariables}
\alias{combineVariables}
\title{Create a data.frame combining set of records from same or multiple variables.}
\usage{
combineVariables(data, paramsList, newVar, labelVars = NULL,
  fctTest = "==", includeAll = FALSE, labelAll = "Any")
}
\arguments{
\item{data}{Data.frame with data.}

\item{paramsList}{nested list of parameters.
There are two ways to select a subset of interest:
\itemize{
\item{by specifying one unique variable of interest with: }{
\itemize{
\item{'var': }{string with column of \code{data} of interest}
\item{'value': }{value of \code{var} of interest (only used if \code{var} is specified)}
\item{'fctTest': }{string with function to apply on \code{var} to select subset of interest 
versus 'value'.
If not specified, the records with \code{var} equal to \code{value} are retained (\code{fctTest} set to '==')
} 
\item{label specification: }{
\itemize{
\item{'label': }{string with complete label for the group}
\item{'labelExtra': }{string with extra label, will be concatenated with 
the label of \code{var} extracted from \code{labelVars}}
}
}
}}
\item{by specifying a combination of variable of interest with: }{
\itemize{
\item{'exprs': }{string with expression of columns of \code{data} to select subset of interest}
\item{'label': }{string with complete label for the group}
}}}}

\item{newVar}{String with name of new variable to construct.}

\item{fctTest}{String with global default function to use to compare \code{var} and \code{value} specified
in each sublist of \code{paramsList}.
This is only used if \code{fctTest} is not specified in each sublist.}

\item{includeAll}{Logical, if TRUE (FALSE by default) include also the entire data as an additional subgroup.}

\item{labelAll}{String of group label for the entire data in case \code{includeAll} is TRUE.}
}
\value{
Data.frame with data and additional variable \code{newVar}.
}
\description{
For each variable, subset of interest can be specified.
}
\author{
Laure Cougnaud
}
