% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineVariables.R
\name{combineVariables}
\alias{combineVariables}
\title{Create a data.frame combining set of records from same or multiple variables.}
\usage{
combineVariables(
  data,
  paramsList,
  newVar,
  labelVars = NULL,
  fctTest = "==",
  includeAll = FALSE,
  labelAll = "Any"
)
}
\arguments{
\item{data}{Data.frame with data.}

\item{newVar}{String with name of new variable to construct.}

\item{labelVars}{Named character vector with variable labels (names are the variable code),
usually extracted from \code{data}.}

\item{fctTest}{Global default function to use to compare \code{var} and \code{value} specified
in each sublist of \code{paramsList}.
This is only used if \code{fctTest} is not specified in each sublist.}

\item{includeAll}{Logical, if TRUE (FALSE by default) include also the entire data as an additional subgroup.}

\item{labelAll}{String of group label for the entire data in case \code{includeAll} is TRUE.}
}
\value{
Data.frame with data and additional variable \code{newVar}.
}
\description{
For each variable, subset of interest can be specified.
}
\author{
Laure Cougnaud
}
