% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSummaryStatistics.R
\name{computeSummaryStatistics}
\alias{computeSummaryStatistics}
\title{Compute summary statistics of interest of an unique variable of interest.}
\usage{
computeSummaryStatistics(
  data,
  var = NULL,
  varTotalInclude = FALSE,
  statsExtra = NULL,
  subjectVar = "USUBJID",
  filterEmptyVar = ((type == "auto" && var != "all" && is.numeric(data[, var])) | type
    == "summaryTable"),
  type = "auto",
  msgLabel = NULL,
  msgVars = NULL
)
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{String, variable of \code{data} with variable to compute statistics on.
Missing values, if present, are filtered.
If NULL or 'all', counts in the entire \code{data} are computed.}

\item{varTotalInclude}{Logical (FALSE by default)
Should the total across all categories of \code{var} 
be included for the count table?
Only used if \code{var} is a categorical variable.}

\item{statsExtra}{(optional) Named list with functions for additional custom
statistics to be computed, only available for 'summaryTable',
e.g. list(statCVPerc = function(x) sd(x)/mean(x)*100) (or \code{\link{cv}}).
Each function has as parameter: either 'x': the variable or 'data': the entire dataset,
and return the corresponding summary statistic.}

\item{subjectVar}{String, variable of \code{data} with subject ID,
'USUBJID' by default.}

\item{filterEmptyVar}{Logical, if TRUE doesn't return any results
if the variable is empty, otherwise return 0 for the counts and NA for summary 
statistics.
Criterias to consider a variable empty are:
\itemize{
\item{for a continuous variable: }{all missing (NA)}
\item{for a categorical variable: }{all missing or **category is included in the 
factor levels but not available in \code{data}**}
}
By default: an empty variable is filtered for a continuous variable
(\code{type} is 'summaryTable' or [\code{type} is 'auto' and 'var' is a numeric variable],
and retained (including all levels) for a categorical variable.}

\item{type}{String with type of table: 
\itemize{
\item{'summaryTable': }{summary table with custom statistics}
\item{'countTable': }{count table}
\item{'auto' (by default): }{'summaryTable' of \code{var} is numeric,
'countTable' otherwise
}
}}

\item{msgLabel}{(optional) String with label for the data (NULL by default), 
included in the message/warning for checks.}

\item{msgVars}{(optional) Character vector with columns of \code{data}
containing extra variables (besides \code{var} and \code{subjectVar})
that should be included in the message/warning for checks.}
}
\value{
Data.frame with summary statistics in columns,
depending if \code{type} is:
\itemize{
\item{'summary': }{
\itemize{
\item{'statN': }{number of subjects}
\item{'statm': }{number of records}
\item{'statMean': }{mean of \code{var}}
\item{'statSD': }{standard deviation of \code{var}}
\item{'statSE': }{standard error the mean of \code{var}}
\item{'statMedian': }{median of \code{var}}
\item{'statMin': }{minimum of \code{var}}
\item{'statMax': }{maximum of \code{var}}
}
}
\item{'count': }{
\itemize{
\item{'variableGroup': }{factor with groups of \code{var} for which counts are reported}
\item{'statN': }{number of subjects}
\item{'statm': }{number of records}
}
}
}
}
\description{
Additionally, this function run extra checks on the data:
\itemize{
\item{an error message is triggered if any subject (identified by \code{subjectVar}) 
have different values in a continuous \code{var}
}
\item{an indicative message is triggered if multiple but identical records are available
for \code{subjectVar} and a continuous \code{var}}
}
}
\author{
Laure Cougnaud
}
