% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSummaryStatistics.R
\name{computeSummaryStatistics}
\alias{computeSummaryStatistics}
\title{Compute summary statistics for a specific dataset and variables of interest}
\usage{
computeSummaryStatistics(data, var = "AVAL", varIgnore = NULL,
  colVar = NULL, rowVar = NULL, subjectVar = "USUBJID",
  type = ifelse(is.numeric(data[, var]), "summaryTable", "countTable"),
  stats = NULL)
}
\arguments{
\item{data}{data.frame with data}

\item{var}{string, variable of \code{data} with variable to use,
'AVAL' by default}

\item{colVar}{variable(s) of \code{data} used 
for grouping in column in the final table. The total 
for each subgroup across \code{rowVar} is computed.}

\item{rowVar}{variable(s) of \code{data} used for
grouping in row in the final table.}

\item{subjectVar}{string, variable of \code{data} with subject ID,
'USUBJID' by default}

\item{type}{string with type of summary table: 'summaryTable' 
(by default if \code{var} is numeric) or 'countTable'}

\item{stats}{(optional) list of expression of summary statistics of interest.
The following statistics are recognized, if: 
\itemize{
\item{\code{type} is a 'summaryTable':}{'N', 'Mean', 'SD', 'SE', 'Median',
'Min', 'Max', 'Perc'}
\item{code{type} is a 'countTable':}{'N','Perc'}
}}
}
\value{
data.frame of class 'countTable' or 'summaryTable',
depending on the 'type' parameter with statistics in columns,
either if \code{type} is:
\itemize{
\item{'summaryTable': }{
\itemize{
\item{'N': }{number of subjects}
\item{'Mean': }{mean of \code{var}}
\item{'SD': }{standard deviation of \code{var}}
\item{'SE': }{standard error of \code{var}}
\item{'Median': }{median of \code{var}}
\item{'Min': }{minimum of \code{var}}
\item{'Max': }{maximum of \code{var}}
\item{'Perc': }{percentage of subjects}
}
}
\item{'countTable': }{
\itemize{
\item{'N': }{number of subjects}
\item{'Perc': }{percentage of subjects}
}
}
}
The computed summary statistics are stored in the 'statsVar' attribute.
}
\description{
Compute summary statistics for a specific dataset and variables of interest
}
\author{
Laure Cougnaud
}
