% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSummaryStatisticsTable.R
\name{computeSummaryStatistics}
\alias{computeSummaryStatistics}
\title{Get summary statistics of interest of an unique variable of interest.}
\usage{
computeSummaryStatistics(data, var = NULL, subjectVar = "USUBJID",
  filterEmptyVar = ((type == "auto" && is.numeric(data[, var])) | type ==
  "summaryTable"), type = "auto")
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{String, variable of \code{data} with variable to compute statistics on.
Missing values, if present, are filtered.}

\item{subjectVar}{String, variable of \code{data} with subject ID,
'USUBJID' by default.}

\item{filterEmptyVar}{Logical, if TRUE doesn't return any results
if the variable is empty, otherwise return 0 for the counts and NA for summary 
statistics (by default, only TRUE if the final table is 'summaryTable')}

\item{type}{String with type of table: 
\itemize{
\item{'summaryTable': }{summary table with custom statistics}
\item{'countTable': }{count table}
\item{'auto' (by default): }{'summaryTable' of \code{var} is numeric,
'countTable' otherwise
}
}}
}
\value{
Data.frame with summary statistics in columns,
depending if \code{type} is:
\itemize{
\item{'summary': }{
\itemize{
\item{'statN': }{number of subjects }
\item{'statm': }{number of records}
\item{'statMean': }{mean of \code{var}}
\item{'statSD': }{standard deviation of \code{var}}
\item{'statSE': }{standard error of \code{var}}
\item{'statMedian': }{median of \code{var}}
\item{'statMin': }{minimum of \code{var}}
\item{'statMax': }{maximum of \code{var}}
}
}
\item{'count': }{
\itemize{
\item{'statN': }{number of subjects}
\item{'statm': }{number of records}
}
}
}
}
\description{
Get summary statistics of interest of an unique variable of interest.
}
\author{
Laure Cougnaud
}
