% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSummaryStatisticsTable.R
\name{computeSummaryStatisticsByRowColVar}
\alias{computeSummaryStatisticsByRowColVar}
\title{Compute summary statistics by specified \code{rowVar} and \code{colVar}}
\usage{
computeSummaryStatisticsByRowColVar(data, var = NULL,
  type = "summaryTable", rowVar = NULL, rowVarInclude0 = FALSE,
  colVar = NULL, colVarInclude0 = TRUE, subjectVar = "USUBJID",
  filterEmptyVar = (type == "summaryTable"), labelVars = NULL)
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{String, variable of \code{data} with variable to compute statistics on.
Missing values, if present, are filtered.}

\item{type}{String with type of summary table: 'summaryTable' 
(by default) or 'countTable'.}

\item{rowVar}{Variable(s) of \code{data} used for
grouping in row in the final table.}

\item{rowVarInclude0}{Logical, if TRUE (FALSE) by default,
include rows with \code{rowVar} elements not available in the data
(e.g. if a \code{rowVar} is a factor, include all levels even if no records in the data).}

\item{colVar}{Variable(s) of \code{data} used 
for grouping in column in the final table. The total 
for each subgroup across \code{rowVar} is computed.}

\item{colVarInclude0}{Logical, if TRUE (FALSE) by default,
include columns with \code{colVar} elements not available in the data
(e.g. if a \code{colVar} is a factor, include all levels even if no records in the data).}

\item{subjectVar}{String, variable of \code{data} with subject ID,
'USUBJID' by default.}

\item{filterEmptyVar}{Logical (TRUE by default), should the summary statistics be filtered
in case \code{var} is empty.}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
data.frame of class 'countTable' or 'summaryTable',
depending on the 'type' parameter; with statistics in columns,
either if \code{type} is:
\itemize{
\item{'summaryTable': }{
\itemize{
\item{'N': }{number of subjects}
\item{'Mean': }{mean of \code{var}}
\item{'SD': }{standard deviation of \code{var}}
\item{'SE': }{standard error of \code{var}}
\item{'Median': }{median of \code{var}}
\item{'Min': }{minimum of \code{var}}
\item{'Max': }{maximum of \code{var}}
\item{'m': }{number of records}
}
}
\item{'countTable': }{
\itemize{
\item{'N': }{number of subjects}
\item{'m': }{number of records}
}
}
}
}
\description{
Compute summary statistics by specified \code{rowVar} and \code{colVar}
}
\author{
Laure Cougnaud
}
