% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSummaryStatisticsTable.R
\name{computeSummaryStatisticsTable}
\alias{computeSummaryStatisticsTable}
\title{Compute summary statistics for a specific dataset and variables of interest}
\usage{
computeSummaryStatisticsTable(data, var = NULL,
  varLab = getLabelVar(var, data = data, labelVars = labelVars),
  varLabGeneral = "Variable", varLabSubgroup = "Subgroup",
  varIgnore = NULL, colVar = NULL, colTotalInclude = FALSE,
  rowVar = NULL, rowVarLab = getLabelVar(rowVar, data = data, labelVars
  = labelVars), rowVarInSepCol = NULL, rowOrder = "auto",
  rowOrderTotalFilterFct = NULL, rowTotalInclude = FALSE,
  rowSubtotalInclude = FALSE, type = "auto", subjectVar = "USUBJID",
  dataTotal = NULL, stats = NULL, filterFct = NULL,
  rowInclude0 = FALSE, colInclude0 = TRUE, labelVars = NULL)
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{String, variable of \code{data} with variable to compute statistics on.
Missing values, if present, are filtered.}

\item{varLab}{Named character vector with label for each variable 
specified in \code{var}.
By default, extracted from the \code{labelVars}.}

\item{varLabGeneral}{String with general label for variable specified in \code{var}.
In case of multiple variable in \code{var}, this will be included in the table header
(see 'rowVarLab' attribute of the output).}

\item{varLabSubgroup}{String with general label for sub-group, in case
\code{var} is specified for a count table.
This will be included in the table header (see 'rowVarLab' attribute of the output).}

\item{varIgnore}{Vector with elements to ignore in the \code{var} variable}

\item{colVar}{Variable(s) of \code{data} used 
for grouping in column in the final table. The total 
for each subgroup across \code{rowVar} is computed.}

\item{colTotalInclude}{Logical, if TRUE (FALSE by default) include the summary 
statistics across columns in a separated column.}

\item{rowVar}{Variable(s) of \code{data} used for
grouping in row in the final table.}

\item{rowVarLab}{Label for the \code{rowVar} variable(s).}

\item{rowVarInSepCol}{Variable(s) of \code{rowVar} which should be 
included in separated column in the table.
NULL by default. 
This is only available if \code{rowVar} if not NULL.}

\item{rowOrder}{String or function of named list with method used to order the rows,
see the \code{method} parameter of \code{\link{convertVarToFactorWithOrder}}.
If a string, the same method is used for all \code{rowVar},
otherwise the list is named with the \code{rowVar} variable, to
specify a different ordering method for each variable.}

\item{rowOrderTotalFilterFct}{Function used to filter the data used to order the rows
based on total counts (in case \code{rowOrder} is 'Total').}

\item{rowTotalInclude}{Logical, if TRUE (FALSE by default) include the total
counts across rows in a separated row.}

\item{type}{String with type of table: 
\itemize{
\item{'summaryTable': }{summary table with custom statistics}
\item{'countTable': }{count table}
\item{'auto' (by default): }{'summaryTable' of \code{var} is numeric,
'countTable' otherwise
}
}}

\item{subjectVar}{String, variable of \code{data} with subject ID,
'USUBJID' by default.}

\item{dataTotal}{Data.frame used to extract the Total count, indicated
in 'N' in column header and used for the computation of the percentage ('Perc') parameter.
It should contain the variables specified by \code{colVar}.}

\item{stats}{(optional) Named list of expression or call object of summary statistics of interest.
The following variables are recognized, if: 
\itemize{
\item{\code{type} is a 'summaryTable':}{'N', 'Mean', 'SD', 'SE', 'Median',
'Min', 'Max', 'Perc'}
\item{\code{type} is a 'countTable':}{'N','Perc'}
}
If \code{stats} if of length 1, the name of the summary statistic is not included
in the table.}

\item{filterFct}{(optional) Function based on computed statistics of
\code{rowVar}/code{colVar} which returns a subset of the summary table of interest.}

\item{rowInclude0}{Logical, if TRUE (FALSE) by default,
include rows with \code{rowVar} elements not available in the data
(e.g. if a \code{rowVar} is a factor, include all levels even if no records in the data).}

\item{colInclude0}{Logical, if TRUE (FALSE) by default,
include columns with \code{colVar} elements not available in the data
(e.g. if a \code{colVar} is a factor, include all levels even if no records in the data).}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
data.frame of class 'countTable' or 'summaryTable',
depending on the 'type' parameter; with statistics in columns,
either if \code{type} is:
\itemize{
\item{'summaryTable': }{
\itemize{
\item{'N': }{number of subjects}
\item{'Mean': }{mean of \code{var}}
\item{'SD': }{standard deviation of \code{var}}
\item{'SE': }{standard error of \code{var}}
\item{'Median': }{median of \code{var}}
\item{'Min': }{minimum of \code{var}}
\item{'Max': }{maximum of \code{var}}
\item{'Perc': }{percentage of subjects}
\item{'m': }{number of records}
\item{'Percm': }{percentage of records}
}
}
\item{'countTable': }{
\itemize{
\item{'N': }{number of subjects}
\item{'PercN': }{percentage of subjects}
\item{'m': }{number of records}
\item{'Percm': }{percentage of records}
}
}
}
The output contains additional the following attributes:
\itemize{
\item{'statsVar': }{column name(s) of summary table with statistics
included in the final table}
\item{'rowVar': column name(s) of summary table with row variable
included in the final table
}
\item{'rowVarLab': labels corresponding to the 'rowVar' attribute
}
}
The computed summary statistics are stored in the 'statsVar' attribute.
}
\description{
Compute summary statistics for a specific dataset and variables of interest
}
\author{
Laure Cougnaud
}
