% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSummaryStatisticsTable.R
\name{computeSummaryStatisticsTable}
\alias{computeSummaryStatisticsTable}
\title{Compute summary statistics for a specific dataset and variables of interest}
\usage{
computeSummaryStatisticsTable(data, var = NULL,
  varLab = getLabelVar(var, data = data, labelVars = labelVars),
  varGeneralLab = "Variable", varSubgroupLab = NULL,
  varIgnore = NULL, varIncludeTotal = FALSE, colVar = NULL,
  colVarDataLevels = NULL, colVarTotal = colVar,
  colVarTotalPerc = colVarTotal, colTotalInclude = FALSE,
  colTotalLab = "Total", rowVar = NULL, rowVarDataLevels = NULL,
  rowVarLab = getLabelVar(rowVar, data = data, labelVars = labelVars),
  rowVarInSepCol = NULL, rowOrder = "auto",
  rowOrderTotalFilterFct = NULL, rowVarTotalInclude = NULL,
  rowVarTotalPerc = NULL, type = "auto", subjectVar = "USUBJID",
  dataTotal = NULL, dataTotalPerc = dataTotal, stats = NULL,
  statsExtra = NULL, statsGeneralLab = "Statistic", filterFct = NULL,
  rowInclude0 = FALSE, colInclude0 = FALSE, labelVars = NULL,
  byVar = NULL, byVarLab = getLabelVar(byVar, data = data, labelVars =
  labelVars))
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{Character vector with variable, used if 'variable' is specified
within \code{rowVarTotal}.}

\item{varLab}{Named character vector with label for each variable 
specified in \code{var}.
By default, extracted from the \code{labelVars}.}

\item{varGeneralLab}{String with general label for variable specified in \code{var}.
In case of multiple variable in \code{var}, this will be included in the table header
(see 'rowVarLab' attribute of the output).}

\item{varSubgroupLab}{String with general label for sub-group, in case
\code{var} is specified for a count table.
This will be included in the table header (see 'rowVarLab' attribute of the output).
Empty by default.}

\item{varIgnore}{Vector with elements to ignore in the \code{var} variable}

\item{varIncludeTotal}{Logical, if TRUE (FALSE by default) the total across 
all categories of \code{var} is included.
Only considered if \code{type} is 'countTable'.}

\item{colVar}{Variable(s) of \code{data} used 
for grouping in column in the final table. The total 
for each subgroup across \code{rowVar} is computed.
If variable(s) are not nested, possible combinations
can be specified via \code{rowVarDataLevels}.}

\item{colVarDataLevels}{Data.frame with levels to consider for the \code{colVar} variable(s)
(in case of non nested column variable(s)).}

\item{colVarTotal}{String with column(s) considered to compute the total by,
reported in the header of the table, by default same as \code{colVar}.}

\item{colVarTotalPerc}{String with column(s) considered to compute the total by,
used as denominator for the percentage computation, by default same as \code{colVarTotal}.}

\item{colTotalInclude}{Logical, if TRUE (FALSE by default) include the summary 
statistics across columns in a separated column.}

\item{colTotalLab}{String, label for the total column included 
if \code{colTotalInclude} is TRUE, 'Total' by default.}

\item{rowVar}{Variable(s) of \code{data} used for
grouping in row in the final table.}

\item{rowVarDataLevels}{Data.frame with levels to consider for the \code{rowVar} variable(s)
(in case of non nested row variable(s)).}

\item{rowVarLab}{Label for the \code{rowVar} variable(s).}

\item{rowVarInSepCol}{Variable(s) of \code{rowVar} which should be 
included in separated column in the table, NULL by default. 
This is only available if \code{rowVar} if not specified.
Note that the row total (if \code{rowTotalInclude} is TRUE) is computed 
separately by this variable.}

\item{rowOrder}{String or function of named list with method used to order the rows,
see the \code{method} parameter of \code{\link{convertVarToFactorWithOrder}}.
'total' is only available if the total is available for each \code{rowVar} (\code{rowSubtotalInclude} set to TRUE)
If a string, the same method is used for all \code{rowVar},
otherwise the list is named with the \code{rowVar} variable, to
specify a different ordering method for each variable.}

\item{rowOrderTotalFilterFct}{Function used to filter the data used to order the rows
based on total counts (in case \code{rowOrder} is 'total').}

\item{rowVarTotalInclude}{Character vector with \code{rowVar}
for which to include the total for each group.
Specify 'all' to include a row with the total across all \code{rowVar}.}

\item{rowVarTotalPerc}{Character vector with row(s) considered to compute the total,
used as denominator for the percentage computation.
By default the total is only computed by column (NULL by default).
If the total should be based on the total number of records per variable,
\code{rowVarTotalPerc} should be set to 'variable'.}

\item{type}{String with type of table: 
\itemize{
\item{'summaryTable': }{summary table with custom statistics}
\item{'countTable': }{count table}
\item{'auto' (by default): }{'summaryTable' of \code{var} is numeric,
'countTable' otherwise
}
}}

\item{subjectVar}{String, variable of \code{data} with subject ID,
'USUBJID' by default.}

\item{dataTotal}{Data.frame used to extract the Total count, indicated
in 'N' in column header.
It should contain the variables specified by \code{colVarTotal}.}

\item{dataTotalPerc}{Data.frame used to extract the Total count, 
used for the computation of the percentage: 'statPercN' parameter.
It should contain the variables specified by \code{colVarTotalPerc}.}

\item{stats}{(Optionally) named list of expression or call object of summary statistics of interest.
The names are reported in the header.
The following variables are recognized, if the table is a: 
\itemize{
\item{'summaryTable': }{
\itemize{
\item{'statN': }{number of subjects}
\item{'statMean': }{mean of \code{var}}
\item{'statSD': }{standard deviation of \code{var}}
\item{'statSE': }{standard error of \code{var}}
\item{'statMedian': }{median of \code{var}}
\item{'statMin': }{minimum of \code{var}}
\item{'statMax': }{maximum of \code{var}}
\item{'statPerc': }{percentage of subjects}
\item{'statm': }{number of records}
}
}
\item{'countTable': }{
\itemize{
\item{'statN': }{number of subjects}
\item{'statPercN': }{percentage of subjects}
\item{'statm': }{number of records}
}
}
}
If \code{stats} if of length 1, the name of the summary statistic is not included
in the table.
The statistics can be specified for each \code{var} (if multiple),
by specifying a nested list of such parameters named by the variable.}

\item{statsExtra}{(optional) Named list with functions for additional custom
statistics to be computed, only available for 'summaryTable',
e.g. list(statCVPercN = function(x) sd(x)/mean(x)*100).
Each function has as parameter: either 'x': the variable or 'data': the entire dataset,
and return the corresponding summary statistic.}

\item{filterFct}{(optional) Function based on computed statistics of
\code{rowVar}/code{colVar} which returns a subset of the summary table 
(after statistics computation).
Note: The filtering function should also handle records with :
\itemize{
\item{\code{rowVar}/\code{codeVar} set to 'Total'/\code{colTotalLab} 
if \code{rowVarTotalInclude}/\code{colTotalInclude} is TRUE}
\item{total for the column header: \code{isTotal} set to TRUE,
and \code{colVar}/\code{rowVar} is NA}
}
Note: the total count per category of the row variables is included
in a row with this variable set to 'Total'.}

\item{rowInclude0}{Logical, if TRUE (FALSE by default),
include rows with no records, based on all combinations 
of the \code{rowVar} (assuming nested variable(s)).}

\item{colInclude0}{Logical, if TRUE (FALSE by default),
include columns with no records, based on all combinations 
of the \code{columnVar} (assuming nested variable(s)).
If variable(s) are not nested, possible combinations
can be specified via \code{colVarDataLevels}.}

\item{labelVars}{named string with variable labels (names are the variable code)}

\item{byVar}{Variable(s) of \code{data} for which separated table(s)
should be created.}

\item{byVarLab}{String with label for \code{byVar}, used to set the names
of the output list of table(s).}
}
\value{
data.frame of class 'countTable' or 'summaryTable',
depending on the 'type' parameter; or list of such data.frame if
\code{byVar} is specified.
The data.frame contains the :
\itemize{
\item{row and column variables}
\item{computed statistics: }{if \code{type} is:
\itemize{
\item{'summaryTable': }{
\itemize{
\item{'statN': }{number of subjects}
\item{'statMean': }{mean of \code{var}}
\item{'statSD': }{standard deviation of \code{var}}
\item{'statSE': }{standard error of \code{var}}
\item{'statMedian': }{median of \code{var}}
\item{'statMin': }{minimum of \code{var}}
\item{'statMax': }{maximum of \code{var}}
\item{'statPerc': }{percentage of subjects}
\item{'statm': }{number of records}
}
}
\item{'countTable': }{
\itemize{
\item{'statN': }{number of subjects}
\item{'statPercN': }{percentage of subjects}
\item{'statm': }{number of records}
}}}
}
\item{statistics specified by \code{statsVar}.
(if not named and of length 1 in a column 'Statistic')}
\item{variables: }{
\itemize{
\item{'variable': }{variable name in case \code{var} is of length > 1}
\item{'variableGroup': }{in case \code{var} is of length > 1 and for variable(s) used for count:
elements of the variable}
}
}
\item{'isTotal': }{variable with logical flag, TRUE if the record contain the total by column}
}
The output contains additional the following attributes:
\itemize{
\item{'statsVar': }{column name(s) of summary table with statistics
included in the final table}
\item{'rowVar': }{column name(s) of summary table with row variable
included in the final table}
\item{'colVar': }{column name(s) of summary table with col variable
included in the final table}
\item{'rowVarLab': }{labels corresponding to the 'rowVar' attribute}
\item{'rowVarInSepCol': }{\code{rowVarInSepCol}}
\item{'rowVarTotalInclude': }{\code{rowVarTotalInclude}}
} 
The computed summary statistics are stored in the 'statsVar' attribute.
}
\description{
Compute summary statistics for a specific dataset and variables of interest
}
\author{
Laure Cougnaud
}
