% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSummaryStatisticsTable.R
\name{computeSummaryStatisticsTable}
\alias{computeSummaryStatisticsTable}
\title{Compute summary statistics for a specific dataset and variables of interest}
\usage{
computeSummaryStatisticsTable(
  data,
  var = NULL,
  varFlag = NULL,
  varInclude0 = FALSE,
  varLab = getLabelVar(var, data = data, labelVars = labelVars),
  varLabInclude = length(var) > 1,
  varGeneralLab = "Variable",
  varSubgroupLab = NULL,
  varIgnore = NULL,
  varIncludeTotal = FALSE,
  varTotalInclude = FALSE,
  varTotalInSepRow = FALSE,
  colVar = NULL,
  colVarDataLevels = NULL,
  colVarTotal = colVar,
  colVarTotalPerc = colVarTotal,
  colTotalInclude = FALSE,
  colTotalLab = "Total",
  rowVar = NULL,
  rowVarDataLevels = NULL,
  rowVarLab = getLabelVar(rowVar, data = data, labelVars = labelVars),
  rowOrder = "auto",
  rowOrderTotalFilterFct = NULL,
  rowOrderCatLast = NULL,
  rowVarTotalInclude = NULL,
  rowVarTotalInSepRow = NULL,
  rowVarTotalByVar = NULL,
  rowVarTotalPerc = NULL,
  type = "auto",
  subjectVar = "USUBJID",
  dataTotal = NULL,
  dataTotalPerc = dataTotal,
  dataTotalRow = NULL,
  dataTotalCol = NULL,
  stats = NULL,
  statsVarBy = NULL,
  statsExtra = NULL,
  statsGeneralLab = "Statistic",
  statsPerc = c("statN", "statm"),
  filterFct = NULL,
  rowInclude0 = FALSE,
  colInclude0 = FALSE,
  labelVars = NULL,
  byVar = NULL,
  byVarLab = getLabelVar(byVar, data = data, labelVars = labelVars)
)
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{Character vector, variable(s) of \code{data} to compute statistics on.
Missing values, if present, are filtered.
If NULL (by default), counts of the \code{rowVar} are returned.
To also return counts of the \code{rowVar} in case other \code{var}
are specified, you can include: 'all' in the \code{var}.}

\item{varFlag}{Character vector, subset of \code{var} with variable(s) 
of type 'flag' (with 'Y', 'N' or '' for empty/non specified value).
Only the counts for records flagged (with 'Y') are retained.}

\item{varInclude0}{Logical, if TRUE (FALSE by default)
include rows with no counts for the count \code{var} or \code{varFlag} variable(s).}

\item{varLab}{Named character vector with label for each variable 
specified in \code{var}.
By default, extracted from the \code{labelVars}.}

\item{varLabInclude}{Logical, if TRUE
the name of the summary statistic variable(s) (\code{var})
are included in the table.
This is automatically set to TRUE if more than one variable(s) 
and is specified, and FALSE if only one variable is specified.}

\item{varGeneralLab}{String with general label for variable specified in \code{var}.
In case of multiple variable in \code{var}, this will be included in the table header
(see 'rowVarLab' attribute of the output).}

\item{varSubgroupLab}{String with general label for sub-group, in case
\code{var} is specified for a count table.
This will be included in the table header (see 'rowVarLab' attribute of the output).
Empty by default.}

\item{varIgnore}{Vector with elements to ignore in the \code{var} variable}

\item{varIncludeTotal}{This argument is deprecated, please use: 'varTotalInclude' instead.}

\item{varTotalInclude}{Should the total across all categories of \code{var} 
be included for the count table?
Only used for categorical variables (and \code{var} not 'all').
Either:
\itemize{
\item{logical of length 1, if TRUE (FALSE by default) include the total for all categorical \code{var}}
\item{a character vector containing categorical \code{var} for which the total should be included}
}}

\item{varTotalInSepRow}{Logical, should the total per variable be included in
a separated row (by default) or in the row containing the header of the variable?}

\item{colVar}{Character vector with variable(s) used for the columns.
If multiple variables are specified, the variables should be sorted in hierarchical order,
and are included in multi-columns layout.
Use: 'variable' to include the variables to summarize: \code{var}
 (if multiple) in different columns.}

\item{colVarDataLevels}{Data.frame with unique combinations of \code{colVar}
to be included in columns.
Each column should correspond to \code{colVar} and as factor
if the elements should be ordered in the final table.}

\item{colVarTotal}{String with column(s) considered to compute the total by,
reported in the header of the table, by default same as \code{colVar}.
Use: 'variable' to compute total by \code{var} (if multiple).}

\item{colVarTotalPerc}{String with column(s) considered to compute the total by,
used as denominator for the percentage computation, by default same as \code{colVarTotal}.
Use: 'variable' to compute total by \code{var} (if multiple).}

\item{colTotalInclude}{Logical, if TRUE (FALSE by default) include the summary 
statistics across columns in a separated column.}

\item{colTotalLab}{String, label for the total column included 
if \code{colTotalInclude} is TRUE, 'Total' by default.}

\item{rowVar}{Variable(s) of \code{data} used for
grouping in row in the final table.}

\item{rowVarDataLevels}{Data.frame with unique combinations of \code{rowVar}
to be included in columns.
Each column should correspond to \code{colVar} and as factor
if the elements should be ordered in the final table.}

\item{rowVarLab}{Label for the \code{rowVar} variable(s).}

\item{rowOrder}{Specify how the rows should be ordered in the table, either a:
\itemize{
\item{String among:}{
\itemize{
\item{'auto' (by default): }{if the variable is a factor, keep its order, otherwise order alphabetically}
\item{'alphabetical': }{order alphabetically}
\item{'total': }{order rows in decreasing order of the total number of subjects
across all columns for this specific category.}
}}
\item{Function with input the summary table and output the ordered elements of the \code{rowVar}}
}
To specify different ordering methods for different \code{rowVar}, specify a list
of such elements, named with the \code{rowVar} variable.}

\item{rowOrderTotalFilterFct}{Function used to filter the data used to order the rows
based on total counts (in case \code{rowOrder} is 'total'),
To order rows based on one specific column category,
e.g. to order based on the counts in the treatment column:
function(x) subset(x, TRTP == "treatmentX")}

\item{rowOrderCatLast}{String with category to be printed in the last 
row of each \code{rowVar} (if any, set to NULL if none).}

\item{rowVarTotalInclude}{Character vector with \code{rowVar}
for which to include the total for each group.}

\item{rowVarTotalInSepRow}{Character vector with \code{rowVarTotalInclude}
(not in \code{rowVarInSepCol}) for which the total should be included in a separated row labelled 'Total'.
Otherwise (by default) the total is included in the header row of each category.}

\item{rowVarTotalByVar}{Character vector with extra \code{rowVar}
for which to compute the row total by.
Can be specified for each \code{rowVarTotalInclude} if named by the corresponding variable.}

\item{rowVarTotalPerc}{Character vector with row variables by which the total
should be computed for the denominator for the percentage computation.
By default the total is only computed by column (NULL by default).
If the total should be based on the total number of records per variable,
\code{rowVarTotalPerc} should be set to 'variable'.}

\item{type}{String with type of table: 
\itemize{
\item{'summaryTable': }{summary table with custom statistics}
\item{'countTable': }{count table}
\item{'auto' (by default): }{'summaryTable' of \code{var} is numeric,
'countTable' otherwise
}
}}

\item{subjectVar}{String, variable of \code{data} with subject ID,
'USUBJID' by default.}

\item{dataTotal}{Data.frame used to extract the Total number of subject
per column in column header ('N = [X]').
It should contain the variables specified by \code{colVarTotal}.
If not specified, the total number of subjects is extracted from the \code{data}.}

\item{dataTotalPerc}{Data.frame used to extract the Total count per column 
for the computation of the percentage.
By default same as \code{dataTotal} .
It should contain the variables specified by \code{colVarTotalPerc}.}

\item{dataTotalRow}{Data.frame used to extract the total count across all
elements of the row
variable, in case \code{rowVarTotalInclude} is specified,
or list of such data.frame for each \code{rowVar} variable (named by variable).}

\item{dataTotalCol}{Data.frame from which the total across columns is 
extracted (in case \code{colTotalInclude} is TRUE)
or list of such data.frame for each \code{rowVar} variable.
This data is used for:
\itemize{
\item{the header of the total column in case \code{dataTotal} is
not specified}
\item{the denominator of the percentages in the total column
in case \code{dataTotalPerc} is not specified}
}}

\item{stats}{(Optionally) Either:
\itemize{
\item{string with: }{
\itemize{
\item{'default': }{default sets of statistics, 
see types: 'summary-default' and 'count-default' in \code{\link{getStats}}}
\item{'all': }{all computed statistics, see types: 'summary' and 'count' in \code{\link{getStats}}}
}}
\item{Named list of expressions or call objects of summary statistics of interest: }{
The names are reported in the header.
The following \strong{'base statistical variables'} are recognized, depending if the variable is: 
\itemize{
\item{continuous (or \code{type} is 'summaryTable'): }{
\itemize{
\item{'statN': }{number of subjects}
\item{'statMean': }{mean of \code{var}}
\item{'statSD': }{standard deviation of \code{var}}
\item{'statSE': }{standard error of \code{var}}
\item{'statMedian': }{median of \code{var}}
\item{'statMin': }{minimum of \code{var}}
\item{'statMax': }{maximum of \code{var}}
\item{'statPercN' (or 'statPercm'): }{percentage of subjects
(or records depending on \code{statsPerc})}
\item{'statm': }{number of records}
}
If multiple and different values are available for a specific \code{var}
and subject ID, by \code{rowVar}/\code{colVar}: an error is triggered.
}
\item{categorical (or \code{type} is 'countTable': }{
\itemize{
\item{'statN': }{number of subjects}
\item{'statPercN' (or 'statPercm'): }{percentage of subjects
(or records depending on \code{statsPerc})}
\item{'statm': }{number of records}
}
}
}
If \code{stats} if of length 1, the name of the summary statistic is not included
in the table.
The statistics can be specified for each \code{var} (if multiple), 
by naming each element of the list:
\code{list(varName1 = list(...), varName2 = list())} and/or for each element in:
\code{statsVarBy}, by naming each sublist.
}}}

\item{statsVarBy}{String with variable in \code{rowVar}/code{colVar}
which the statistics should be computed by.
In this case, \code{stats} (nested list or not) should be additionally nested
to specify the statistics for each element in \code{statsVarBy}.}

\item{statsExtra}{(optional) Named list with functions for additional custom
statistics to be computed, only available for 'summaryTable',
e.g. list(statCVPerc = function(x) sd(x)/mean(x)*100) (or \code{\link{cv}}).
Each function has as parameter: either 'x': the variable or 'data': the entire dataset,
and return the corresponding summary statistic.}

\item{statsGeneralLab}{String with general label for statistics, 'Statistic' by default.
Only included if no \code{statsVar} if longer than 1.}

\item{statsPerc}{String with 'base statistical variable' used to compute the 
percentage, either: 
\itemize{
\item{'statN' (by default): }{the number of subjects}
\item{'statm': }{the number of records}
}}

\item{filterFct}{(optional) Function taking as input
the summary table with computed statistics and returning a subset 
of the summary table.\cr
Note: The filtering function should also handle records with :
\itemize{
\item{total for the column header: \code{isTotal} set to TRUE,
and \code{colVar}/\code{rowVar} is NA.\cr
For example: \code{filterFct = function(data) subset(data, isTotal & myColVar == "group 1")}
}
\item{\code{rowVar}/\code{colVar} set to 'Total'/\code{colTotalLab} 
if \code{rowVarTotalInclude}/\code{colTotalInclude} is specified}
}}

\item{rowInclude0}{Logical, if TRUE (FALSE by default),
include rows with no records, based on all combinations 
of the \code{rowVar} (assuming nested variable(s)).}

\item{colInclude0}{Logical, if TRUE (FALSE by default),
include columns with no records, based on all combinations 
of the \code{columnVar} (assuming nested variable(s)).
If variable(s) are not nested, possible combinations
can be specified via \code{colVarDataLevels}.}

\item{labelVars}{Named character vector with variable labels (names are the variable code),
usually extracted from \code{data}.}

\item{byVar}{Variable(s) of \code{data} for which separated table(s)
should be created.}

\item{byVarLab}{String with label for \code{byVar}, used to set the names
of the output list of table(s).}
}
\value{
data.frame of class 'countTable' or 'summaryTable',
depending on the 'type' parameter; or list of such data.frame if
\code{byVar} is specified.
The data.frame contains the :
\itemize{
\item{row and column variables}
\item{computed statistics: }{if \code{type} is:
\itemize{
\item{'summaryTable': }{
\itemize{
\item{'statN': }{number of subjects}
\item{'statMean': }{mean of \code{var}}
\item{'statSD': }{standard deviation of \code{var}}
\item{'statSE': }{standard error of \code{var}}
\item{'statMedian': }{median of \code{var}}
\item{'statMin': }{minimum of \code{var}}
\item{'statMax': }{maximum of \code{var}}
\item{'statPerc': }{percentage of subjects}
\item{'statPercTotalN': }{total number of subjects based on \code{dataTotalPerc},
denominator of \code{statPerc}}
\item{'statm': }{number of records}
}
}
\item{'countTable': }{
\itemize{
\item{'statN': }{number of subjects}
\item{'statPercN' (or 'statPercm'): }{percentage of subjects
(or records depending on \code{statsPerc})}
\item{'statPercTotalN' (or 'statPercTotalm'): }{total number of 
subjects (or records) based on \code{dataTotalPerc}, and used as
denominator of \code{statPercN} (or 'statPercm')}
\item{'statm': }{number of records}
}}}
}
\item{statistics specified by \code{statsVar}.
(if not named and of length 1 in a column 'Statistic')}
\item{variables: }{
\itemize{
\item{'variable': }{variable name in case \code{var} is of length > 1}
\item{'variableGroup': }{in case \code{var} is of length > 1 and for variable(s) used for count:
elements of the variable}
}
}
\item{'isTotal': }{variable with logical flag, TRUE if the record contain the total by column}
}
Additionally, the output contains the following attributes:
\itemize{
\item{'statsVar': }{column name(s) of summary table with computed statistics
included in the final table}
\item{'rowVar': }{column name(s) of summary table with row variable
included in the final table. This parameter should be mainly used for qualitative variables and 'nests' together different rows in the final output table.}
\item{'rowVarLab': }{labels corresponding to the 'rowVar' attribute}
\item{'rowVarTotalInclude': }{row variables whose total will be included:
\code{rowVarTotalInclude} and 'variableGroup' if the variable total should be included}
\item{'rowVarTotalInSepRow': }{row variables whose total will be included in a separated row:
\code{rowVarTotalInSepRow} and 'variableGroup' if \code{varTotalInSepRow}}
\item{'colVar': }{column name(s) of summary table with column variable
included in the final table}
\item{'colTotalLab': }{label for the total}
}
}
\description{
Compute summary statistics for a specific dataset and variables of interest
}
\author{
Laure Cougnaud
}
