% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSummaryStatisticsTable.R
\name{computeSummaryStatisticsTableTotal}
\alias{computeSummaryStatisticsTableTotal}
\title{Compute summary statistics total table.}
\usage{
computeSummaryStatisticsTableTotal(
  data,
  colVar = NULL,
  colVarTotal = colVar,
  rowVarTotal = NULL,
  var = NULL,
  varLab = getLabelVar(var, data = data, labelVars = labelVars),
  colTotalLab = "Total",
  colInclude0 = FALSE,
  colTotalInclude = FALSE,
  dataTotalCol = NULL,
  colVarDataLevels = NULL,
  colVarLevels = NULL,
  subjectVar = "USUBJID",
  labelVars = NULL,
  msgLabel = "total"
)
}
\arguments{
\item{data}{Data.frame with data.}

\item{colVar}{Character vector with variable(s) used for the columns.
If multiple variables are specified, the variables should be sorted in hierarchical order,
and are included in multi-columns layout.
Use: 'variable' to include the variables to summarize: \code{var}
 (if multiple) in different columns.}

\item{colVarTotal}{Character vector with column(s) considered to compute the total.
This could also contain 'variable'.}

\item{rowVarTotal}{Character vector with row(s) considered to compute the total.
This could also contain 'variable'.}

\item{var}{Character vector with variable, used if 'variable' is specified
within \code{rowVarTotal}.}

\item{varLab}{Named character vector with label for each variable 
specified in \code{var}.
By default, extracted from the \code{labelVars}.
if not available, \code{var} is used.}

\item{colInclude0}{Logical, if TRUE (FALSE by default),
include columns with no records, based on all combinations 
of the \code{columnVar} (assuming nested variable(s)).
If variable(s) are not nested, possible combinations
can be specified via \code{colVarDataLevels}.}

\item{colTotalInclude}{Logical, if TRUE (FALSE by default) include the summary 
statistics across columns in a separated column.}

\item{dataTotalCol}{Data.frame from which the total across columns is 
extracted (in case \code{colTotalInclude} is TRUE).}

\item{colVarDataLevels}{Data.frame with unique combinations of \code{colVar}
to be included in columns.
Each column should correspond to \code{colVar} and as factor
if the elements should be ordered in the final table.}

\item{colVarLevels}{list with levels of each \code{colVar}}

\item{subjectVar}{String, variable of \code{data} with subject ID,
'USUBJID' by default.}

\item{labelVars}{Named character vector with variable labels (names are the variable code),
usually extracted from \code{data}.}

\item{msgLabel}{(optional) String with label for the data ('total' by default), 
included in the message/warning for checks.}
}
\value{
data.frame with total table.
}
\description{
Compute summary statistics total table.
}
\author{
Laure Cougnaud
}
\keyword{internal}
