% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSummaryStatisticsTable.R
\name{computeSummaryStatisticsTableTotal}
\alias{computeSummaryStatisticsTableTotal}
\title{Compute summary statistics total table.}
\usage{
computeSummaryStatisticsTableTotal(data, colVar = NULL,
  colVarTotal = colVar, rowVarTotal = NULL, var = NULL,
  varLab = getLabelVar(var, data = data, labelVars = labelVars),
  colTotalLab = "Total", colInclude0 = FALSE,
  colTotalInclude = FALSE, colVarDataLevels = NULL,
  colVarLevels = NULL, subjectVar = "USUBJID", labelVars = NULL)
}
\arguments{
\item{data}{Data.frame with data.}

\item{colVar}{Variable(s) of \code{data} used 
for grouping in column in the final table. The total 
for each subgroup across \code{rowVar} is computed.
If variable(s) are not nested, possible combinations
can be specified via \code{rowVarDataLevels}.}

\item{colVarTotal}{Character vector with column(s) considered to compute the total.}

\item{rowVarTotal}{Character vector with row(s) considered to compute the total.}

\item{var}{Character vector with variable, used if 'variable' is specified
within \code{rowVarTotal}.}

\item{varLab}{Named character vector with label for each variable 
specified in \code{var}.
By default, extracted from the \code{labelVars}.}

\item{colTotalLab}{String, label for the total column included 
if \code{colTotalInclude} is TRUE, 'Total' by default.}

\item{colInclude0}{Logical, if TRUE (FALSE by default),
include columns with no records, based on all combinations 
of the \code{columnVar} (assuming nested variable(s)).
If variable(s) are not nested, possible combinations
can be specified via \code{colVarDataLevels}.}

\item{colTotalInclude}{Logical, if TRUE (FALSE by default) include the summary 
statistics across columns in a separated column.}

\item{colVarDataLevels}{Data.frame with levels to consider for the \code{colVar} variable(s)
(in case of non nested column variable(s)).}

\item{colVarLevels}{list with levels of each \code{colVar}}

\item{subjectVar}{String, variable of \code{data} with subject ID,
'USUBJID' by default.}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
data.frame with total table.
}
\description{
Compute summary statistics total table.
}
\author{
Laure Cougnaud
}
