% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSummaryStatisticsTable.R
\name{convertSummaryStatisticsTableToFlextable}
\alias{convertSummaryStatisticsTableToFlextable}
\title{Convert summary statistics table to flextable}
\usage{
convertSummaryStatisticsTableToFlextable(summaryTable, landscape = FALSE,
  margin = 1, rowPadBase = 2,
  title = "Table: Descriptive statistics", footer = NULL,
  fontname = "Times")
}
\arguments{
\item{summaryTable}{summary statistics table in long format,
as returned by \code{\link{formatSummaryStatisticsForExport}}}

\item{landscape}{logical, if TRUE (by defaut) the table is presented in landscape
format}

\item{margin}{margin in the document in inches}

\item{rowPadBase}{base padding for row (number of spaces)}

\item{title}{character vector with title(s) for the table.
Set to NULL if no title should be included.}

\item{footer}{character vector with footer(s) for the table.
Set to NULL (by default) of no footer should be included.}

\item{fontname}{string with font name, 'Times' by default}
}
\value{
\code{\link[flextable]{flextable}} object with summary table
}
\description{
Convert summary statistics table to flextable
}
\author{
Laure Cougnaud
}
