% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSummaryStatisticsTableToFlextable.R
\name{convertSummaryStatisticsTableToFlextable}
\alias{convertSummaryStatisticsTableToFlextable}
\title{Convert summary statistics table to flextable}
\usage{
convertSummaryStatisticsTableToFlextable(summaryTable, landscape = (style
  == "presentation"), margin = 1, rowPadBase = 14.4, title = NULL,
  footer = NULL, style = "report", fontname = switch(style, report =
  "Times", presentation = "Tahoma"), fontsize = switch(style, report = 8,
  presentation = 10), file = NULL)
}
\arguments{
\item{summaryTable}{Summary table, created with the \code{\link{computeSummaryStatisticsTable}} function.}

\item{landscape}{Logical, if TRUE (by defaut) the table is presented in landscape
format.}

\item{margin}{Margin in the document in inches.}

\item{rowPadBase}{Base padding for row (in points)}

\item{title}{Character vector with title(s) for the table.
Set to NULL (by default) if no title should be included.}

\item{footer}{Character vector with footer(s) for the table.
Set to NULL (by default) of no footer should be included.}

\item{style}{string with table style, either 'report' or 'presentation'}

\item{fontname}{String with font name, 'Times' by default.}

\item{fontsize}{Integer with font size, 8 by default.}

\item{file}{String with path of the file where the table should be exported.
If NULL, the summary table is not exported but only returned as output.}
}
\value{
\code{\link[flextable]{flextable}} object with summary table
If \code{summaryTable} is a list of summary tables,
returns a list of \code{\link[flextable]{flextable}}.
}
\description{
Convert summary statistics table to flextable
}
\author{
Laure Cougnaud
}
