% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSummaryStatisticsTableToFlextable.R
\name{convertSummaryStatisticsTableToFlextable}
\alias{convertSummaryStatisticsTableToFlextable}
\title{Convert summary statistics table to flextable}
\usage{
convertSummaryStatisticsTableToFlextable(summaryTable, landscape = FALSE,
  margin = 1, rowPadBase = 4, title = NULL, footer = NULL,
  fontname = "Times")
}
\arguments{
\item{summaryTable}{Dummary statistics table in long format,
as returned by \code{\link{formatSummaryStatisticsTable}}.}

\item{landscape}{Logical, if TRUE (by defaut) the table is presented in landscape
format.}

\item{margin}{Margin in the document in inches.}

\item{rowPadBase}{Base padding for row (number of spaces)}

\item{title}{Character vector with title(s) for the table.
Set to NULL (by default) if no title should be included.}

\item{footer}{Character vector with footer(s) for the table.
Set to NULL (by default) of no footer should be included.}

\item{fontname}{String with font name, 'Times' by default}
}
\value{
\code{\link[flextable]{flextable}} object with summary table
If \code{summaryTable} is a list of summary tables,
returns a list of \code{\link[flextable]{flextable}}.
}
\description{
Convert summary statistics table to flextable
}
\author{
Laure Cougnaud
}
