% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable-getListing.R
\name{convertVectToBinary}
\alias{convertVectToBinary}
\title{Convert vector to a bincode of 0/1
based on consecutive values in the vector.}
\usage{
convertVectToBinary(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
Integer vector of same length than \code{x}.
}
\description{
Convert vector to a bincode of 0/1
based on consecutive values in the vector.
}
\examples{
x <- c("group1", "group1", "group1", "group2", "group2", "group3", "group4", "group4")
convertVectToBinary(x = x)
}
\author{
Laure Cougnaud
}
