% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSummaryStatisticsTableToFlextable.R
\name{createFlextableWithHeader}
\alias{createFlextableWithHeader}
\title{Create a flextable, setting the column names to syntactic names
if it is not the case.}
\usage{
createFlextableWithHeader(data, headerDf = NULL)
}
\arguments{
\item{data}{Data.frame with data.}

\item{headerDf}{(optional) Data.frame with header.}
}
\value{
list with:
\itemize{
\item{'ft': }{\code{\link[flextable]{flextable}}}
\item{'colsData': }{Named vector with original column names,
with names set to new syntactic names.}
}
}
\description{
Create a flextable, setting the column names to syntactic names
if it is not the case.
}
\author{
Laure Cougnaud
}
