% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSummaryStatisticsTable.R
\name{exportSummaryStatisticsTable}
\alias{exportSummaryStatisticsTable}
\title{Export a summary table in \code{docx} format.}
\usage{
exportSummaryStatisticsTable(summaryTable, rowVar = NULL,
  rowVarLab = getLabelVar(rowVar, labelVars = labelVars),
  rowVarInCol = NULL, colVar = NULL, labelVars = NULL, file = NULL,
  landscape = FALSE, margin = 1, rowPadBase = 2,
  title = "Table: Descriptive statistics")
}
\arguments{
\item{summaryTable}{summary statistics table in long format,
as returned by \code{\link{formatSummaryStatisticsForExport}}}

\item{rowVar}{character vector with variable(s) of \code{summaryTable}
used for grouping in rows.}

\item{rowVarLab}{label for each variable of \code{rowVar}.}

\item{colVar}{string with variable of \code{summaryTable} used for grouping in column.}

\item{labelVars}{named string with variable labels (names are the variable code)}

\item{file}{string with path of the file where the table should be exported}

\item{landscape}{logical, if TRUE (by defaut) the table is presented in landscape
format}

\item{margin}{margin in the document in inches}

\item{rowPadBase}{base padding for row (number of spaces)}

\item{title}{string with title for the table.
Set to NULL if no title should be included.}
}
\description{
Export a summary table in \code{docx} format.
}
\author{
Laure Cougnaud
}
