% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSummaryStatisticsTable.R
\name{exportSummaryStatisticsTable}
\alias{exportSummaryStatisticsTable}
\title{Export a summary table in \code{docx} format.}
\usage{
exportSummaryStatisticsTable(summaryTable, rowVar = NULL,
  rowVarLab = getLabelVar(rowVar, labelVars = labelVars),
  rowVarInSepCol = NULL, rowTotalInclude = FALSE, rowTotalLab = NULL,
  colVar = NULL, labelVars = NULL, file = NULL, landscape = FALSE,
  margin = 1, rowPadBase = 2,
  title = "Table: Descriptive statistics", footer = NULL)
}
\arguments{
\item{summaryTable}{summary statistics table in long format,
as returned by \code{\link{formatSummaryStatisticsForExport}}}

\item{rowVar}{character vector with variable(s) used for the rows.
If multiple variables are specified, the variables should be sorted in hierarchical order.
The variables are included in rows, excepted if specified in \code{rowVarInSepCol}.}

\item{rowVarLab}{label for the \code{rowVar} variable(s)}

\item{rowVarInSepCol}{variable(s) of \code{rowVar} which should be 
included in separated column in the table.
NULL by default. 
This is only available if \code{rowVar} if not NULL.}

\item{rowTotalInclude}{logical, if TRUE (FALSE by default) include the total
across rows in a separated row.}

\item{rowTotalLab}{label for the row with total}

\item{colVar}{character vector with variable(s) used for the columns.
If multiple variables are specified, the variables should be sorted in hierarchical order,
and are included in multi-columns layout.}

\item{labelVars}{named string with variable labels (names are the variable code)}

\item{file}{string with path of the file where the table should be exported.
If NULL, the summary table is not exported but only returned as output.}

\item{landscape}{logical, if TRUE (by defaut) the table is presented in landscape
format}

\item{margin}{margin in the document in inches}

\item{rowPadBase}{base padding for row (number of spaces)}

\item{title}{character vector with title(s) for the table.
Set to NULL if no title should be included.}

\item{footer}{character vector with footer(s) for the table.
Set to NULL (by default) of no footer should be included.}
}
\value{
\code{\link[flextable]{flextable}} object with summary table
}
\description{
Export a summary table in \code{docx} format.
}
\author{
Laure Cougnaud
}
