% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSummaryStatisticsTable.R
\name{exportSummaryStatisticsTable}
\alias{exportSummaryStatisticsTable}
\title{Export a summary table in \code{docx} format.}
\usage{
exportSummaryStatisticsTable(
  summaryTable,
  rowVar = getAttribute(summaryTable, "rowVar"),
  rowVarLab = getAttribute(summaryTable, "rowVarLab", default = getLabelVar(rowVar,
    labelVars = labelVars)),
  rowVarInSepCol = NULL,
  rowVarFormat = NULL,
  rowVarTotalInclude = getAttribute(summaryTable, "rowVarTotalInclude"),
  rowTotalLab = NULL,
  rowVarTotalInSepRow = getAttribute(summaryTable, "rowVarTotalInSepRow"),
  rowAutoMerge = TRUE,
  colVar = getAttribute(summaryTable, "colVar"),
  colTotalLab = getAttribute(summaryTable, "colTotalLab", default = "Total"),
  colHeaderTotalInclude = TRUE,
  statsVar = getAttribute(summaryTable, "statsVar"),
  statsLayout = getAttribute(summaryTable, "statsLayout", default = ifelse("DT" \%in\%
    outputType, "col", "row")),
  statsValueLab = "StatisticValue",
  statsLabInclude = NULL,
  emptyValue = "-",
  labelVars = NULL,
  file = NULL,
  landscape = (style == "presentation"),
  margin = 1,
  rowPadBase = 14.4,
  title = NULL,
  outputType = "flextable",
  footer = NULL,
  style = "report",
  colorTable = getColorTable(style = style),
  fontsize = switch(style, report = 8, presentation = 10),
  fontname = switch(style, report = "Times", presentation = "Tahoma"),
  vline = "none",
  hline = "auto",
  pageDim = NULL,
  expandVar = NULL,
  noEscapeVar = NULL,
  barVar = NULL,
  ...
)
}
\arguments{
\item{summaryTable}{Summary table, created with the \code{\link{computeSummaryStatisticsTable}} function.}

\item{rowVar}{Character vector with variable(s) used for the rows.
If multiple variables are specified, the variables should be sorted in hierarchical order.
The variables are included in rows, excepted if specified in \code{rowVarInSepCol}.}

\item{rowVarLab}{Label for the \code{rowVar} variable(s).}

\item{rowVarInSepCol}{Character vector with \code{rowVar}
that should be included in multiple columns.
NULL by default: the row variables are nested in the first column 
of the table.
To include the groups within a \code{var} variable in a separated column, 
set: rowVarInSepCol == 'variableGroup'.
This is only available if \code{rowVar} is not specified.}

\item{rowVarFormat}{Named list with special formatting for the \code{rowVar}.
Currently, only possibility is to set the variable elements in bold, with:
list(var1 = "bold").
(Use 'variable' for \code{var} or 'variableGroup' for group within categorical variables.)}

\item{rowVarTotalInclude}{Character vector with \code{rowVar} for which the total
is included. These variables should be formatted as a factor with 
\strong{'Total' as the first level}.}

\item{rowTotalLab}{label for the row with total}

\item{rowVarTotalInSepRow}{Character vector with \code{rowVarTotalInclude}
(not in \code{rowVarInSepCol}) for which the total should be included in a separated row labelled 'Total'.
Otherwise (by default) the total is included in the header row of each category.}

\item{rowAutoMerge}{Logical, if TRUE (by default) automatically merge rows,
e.g. in case there is only one sub-category (e.g. categorical variable with only one group)
or only one statistic per category.}

\item{colVar}{Character vector with variable(s) used for the columns.
If multiple variables are specified, the variables should be sorted in hierarchical order,
and are included in multi-columns layout.
Use: 'variable' to include the variables to summarize: \code{var}
 (if multiple) in different columns.}

\item{colTotalLab}{String, label for the total column included 
if \code{colTotalInclude} is TRUE, 'Total' by default.}

\item{colHeaderTotalInclude}{Logical, if TRUE include the total of number of patients
(\code{'statN'}) in the header.}

\item{statsVar}{Character vector with columns of \code{summaryTable} with
statistic variables.}

\item{statsLayout}{String with layout for the statistics names 
(in case more than one statistic is included), among:
\itemize{
\item{row (by default when \code{outputType} is: 'flextable'): }{
All statistics are included in different rows 
in the first column of the table
}
\item{'col' (by default when \code{outputType} is: 'DT'): }{
Statistics are included in separated columns (last row of the header).
This option is not compatible with categorical variable(s).
}
\item{'rowInSepCol': }{
Statistics are included in different rows, but in a separated column than
the \code{rowVar} variable(s)
}
}}

\item{statsValueLab}{String with label for the statistic value, 
'StatisticValue' by default.
This is only used if the statistics provided in \code{stats} are not named
and if no \code{colVar} is specified.}

\item{statsLabInclude}{Logical, if TRUE include the statistic label
in the table. By default only included if more than
one statistic variables are available in the table.}

\item{emptyValue}{String with placeholder used to fill the table for missing values, '-' by default.
This value is typically used e.g. if not all statistics are computed for all specified
row/col/var variables.}

\item{labelVars}{Named character vector with variable labels (names are the variable code),
usually extracted from \code{data}.}

\item{file}{(Optional) Filename to which to which the table should be exported, 
either:
\itemize{
\item{string (of length 1). In this case, depending on the
file extension, the following is exported: }{
\itemize{
\item{'txt': summary table in long format ('data.frame-base' \code{outputType})}
\item{'docx': summary table in final format is exported ('flextable' \code{outputType})}
\item{'html': interactive summary table is exported ('DT' \code{outputType})}
}}
\item{named character vector in case of multiple exports.
The names should correspond to the options in \code{outputType}:}{
\itemize{
\item{for 'data.frame-base' and 'data.frame': }{filename with 'txt' extension}
\item{for 'flextable': }{filename with 'docx' extension}
\item{for 'DT': }{filename with 'html' extension}
}}
}}

\item{landscape}{Logical, if TRUE the table is presented in landscape
format. \cr
By default: TRUE for \code{style}: 'report',
FALSE for \code{style}: 'presentation'.}

\item{margin}{Margin in the document in inches, currently only used to specify the
width of the table: [width page extracted from \code{dimPage} - 2* margin]}

\item{rowPadBase}{Base padding for row (in points), 14.4 by default (corresponds to 0.2 inches)}

\item{title}{Character vector with title(s) for the table.
Set to NULL (by default) if no title should be included.
If vector > 1, specified for each element of \code{byVar} (in order of the levels).}

\item{outputType}{String with output type:
\itemize{
\item{'flextable' (by default): }{\code{\link[flextable]{flextable}} object, with format for
CSR, compatible with Word/PowerPoint export}
\item{'DT'}{interactive table: \code{\link[DT]{datatable}} ,
e.g. to be exported to html}
\item{'data.frame': }{data.frame in wide format (with elements in 
\code{colVar} in different columns)}
\item{'data.frame-base'}{data.frame in long format (with elements in 
\code{colVar} in different rows), useful for debugging}
}}

\item{footer}{Character vector with footer(s) for the table.
Set to NULL (by default) of no footer should be included.}

\item{style}{string with table style in case \code{outputType} is 'flextable',
 either 'report' or 'presentation'.
This parameter affects the fontsize, font family, color of the text and background, 
and table dimensions of the table.}

\item{colorTable}{Named character vector with color for the table background/body/text/line,
e.g. created with the \code{\link[glpgStyle]{getColorTable}} function.}

\item{fontsize}{Integer with font size, by default:
8 if \code{style} is 'report' and 10 if \code{style} is 'presentation'.}

\item{fontname}{String with font name, by default:
'Times' if \code{style} is 'report' and 'Tahoma' if \code{style} is 'presentation'.}

\item{vline}{String mentioning how vertical lines 
should be included in the body of the table, either: 
\itemize{
\item{'none' (default): }{no vertical lines included}
\item{'auto': }{vertical lines included between sub-groups}
}}

\item{hline}{String mentioning how horizontal lines 
should be included in the body of the table, either: 
\itemize{
\item{'none': }{no horizontal lines included}
\item{'auto' (default): }{horizontal lines included between sub-groups}
}}

\item{pageDim}{Numeric vector of length 2 with page width and height
depending on \code{outputType}:
\itemize{
\item{'flextable': }{in inches in portrait format}
\item{'DT': }{number of rows in the table (currently only
the height is used (e.g. \code{c(NA, 4)})
}}}

\item{expandVar}{Character vector with variables of the summary table which
should be expanded in the data (only for 'DT' output).}

\item{noEscapeVar}{Character vector with variables of \code{summaryTable}
which shouldn't be escaped in the table (e.g. containing URLs).}

\item{barVar}{Character vector with variables of \code{summaryTable}
that should be represented as a bar.}

\item{...}{Extra parameters passed to the 
\code{\link[glpgUtilityFct]{toDTGLPG}} (only for 'DT' output)}
}
\value{
Depending on the \code{outputType}:
\itemize{
\item{'data.frame-base': }{input summary table in a long format with
all computed statistics}
\item{'data.frame': }{summary table in a wide format (
different columns for each \code{colVar}), with specified labels}
\item{'flextable' (by default): }{\code{\link[flextable]{flextable}}
 object with summary table}
\item{'DT': }{\code{\link[DT]{datatable}} object with summary table}
}
If \code{summaryTable} is a list of summary tables,
returns a list of corresponding summary tables in long format.
}
\description{
Export a summary table in \code{docx} format.
}
\author{
Laure Cougnaud
}
