% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSummaryStatisticsTable.R
\name{formatSummaryStatisticsForExport}
\alias{formatSummaryStatisticsForExport}
\title{Format summary statistics table for export}
\usage{
formatSummaryStatisticsForExport(summaryTable, rowVar = NULL,
  rowVarLab = getLabelVar(rowVar, labelVars = labelVars),
  rowVarInCol = NULL, colVar = NULL, labelVars = NULL)
}
\arguments{
\item{rowVar}{character vector with variable(s) of \code{summaryTable}
used for grouping in rows.}

\item{rowVarLab}{label for each variable of \code{rowVar}.}

\item{colVar}{string with variable of \code{summaryTable} used for grouping in column.}

\item{labelVars}{named string with variable labels (names are the variable code)}
}
\value{
summaryTable reformatted in long format, with extra attributes:
\itemize{
\item{'header': }{data.frame with header for each column}
\item{'padParams': }{list of list of parameters to be passed to the 
\code{\link[flextable]{padding}}} function
}
}
\description{
Format summary statistics table for export
}
\author{
Laure Cougnaud
}
