% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatSummaryStatisticsForExport.R
\name{formatSummaryStatisticsForExport}
\alias{formatSummaryStatisticsForExport}
\title{Format summary statistics table for export}
\usage{
formatSummaryStatisticsForExport(summaryTable, rowVar = NULL,
  rowVarLab = getLabelVar(rowVar, labelVars = labelVars),
  rowVarInSepCol = NULL, rowTotalInclude = FALSE, rowTotalLab = NULL,
  colVar = NULL, labelVars = NULL)
}
\arguments{
\item{summaryTable}{Summary table, created with the \code{\link{computeSummaryStatistics}} function.}

\item{rowVar}{Character vector with variable(s) used for the rows.
If multiple variables are specified, the variables should be sorted in hierarchical order.
The variables are included in rows, excepted if specified in \code{rowVarInSepCol}.}

\item{rowVarLab}{Label for the \code{rowVar} variable(s).}

\item{rowVarInSepCol}{Variable(s) of \code{rowVar} which should be 
included in separated column in the table.
NULL by default. 
This is only available if \code{rowVar} if not NULL.}

\item{rowTotalInclude}{Logical, if TRUE (FALSE by default) include the total
across rows in a separated row.}

\item{rowTotalLab}{label for the row with total}

\item{colVar}{Character vector with variable(s) used for the columns.
If multiple variables are specified, the variables should be sorted in hierarchical order,
and are included in multi-columns layout.}

\item{labelVars}{Named string with variable labels (names are the variable code).}
}
\value{
summaryTable reformatted in long format, with extra attributes:
\itemize{
\item{'header': }{data.frame with header for each column}
\item{'padParams': }{list of list of parameters to be passed to the 
\code{\link[flextable]{padding}} function}
\item{'rowVar': }{column of output with row variable}
\item{'vlineParams' and 'hlineParams': }{
list of list with correspondingly parameters for
vertical and horizontal lines}
}
}
\description{
Format summary statistics table for export
}
\author{
Laure Cougnaud
}
