% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatSummaryStatisticsTable.R
\name{formatSummaryStatisticsTable}
\alias{formatSummaryStatisticsTable}
\title{Format summary statistics table for export}
\usage{
formatSummaryStatisticsTable(
  summaryTable,
  rowVar = getAttribute(summaryTable, "rowVar"),
  colVar = getAttribute(summaryTable, "colVar"),
  colTotalLab = getAttribute(summaryTable, "colTotalLab", default = "Total"),
  colHeaderTotalInclude = TRUE,
  statsVar = getAttribute(summaryTable, "statsVar"),
  statsLayout = "row",
  statsValueLab = "StatisticValue",
  emptyValue = "-"
)
}
\arguments{
\item{summaryTable}{Summary table, created with the \code{\link{computeSummaryStatisticsTable}} function.}

\item{rowVar}{Character vector with variable(s) used for the rows.
If multiple variables are specified, the variables should be sorted in hierarchical order.
The variables are included in rows, excepted if specified in \code{rowVarInSepCol}.}

\item{colVar}{Character vector with variable(s) used for the columns.
If multiple variables are specified, the variables should be sorted in hierarchical order,
and are included in multi-columns layout.}

\item{colTotalLab}{String, label for the total column included 
if \code{colTotalInclude} is TRUE, 'Total' by default.}

\item{colHeaderTotalInclude}{Logical, if TRUE include the total of number of patients
(\code{'statN'}) in the header.}

\item{statsVar}{Character vector with columns of \code{summaryTable} with
statistic variables.}

\item{statsLayout}{String with layout for the statistics names 
(in case more than one statistic is included), among:
\itemize{
\item{row (by default): }{All statistics are included in different rows 
in the first column of the table}
\item{'col': }{Statistics are included in separated columns (last row of the header).
This option is not compatible with categorical variable(s).}
\item{'rowInSepCol': }{Statistics are included in different rows, but in a separated column than
the \code{rowVar} variable(s)}
}}

\item{statsValueLab}{String with label for the statistic value, 
'StatisticValue' by default.
This is only used if the statistics provided in \code{stats} are not named
and if no \code{colVar} is specified.}

\item{emptyValue}{Value used to fill the table for missing values, '-' by default.
See the \code{fill} parameter of the \code{\link[reshape2]{dcast}} function.}
}
\value{
summaryTable reformatted to wide format
}
\description{
Format summary statistics table for export
}
\author{
Laure Cougnaud
}
