% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatSummaryStatisticsTable.R
\name{formatSummaryStatisticsTable}
\alias{formatSummaryStatisticsTable}
\title{Format summary statistics table for export}
\usage{
formatSummaryStatisticsTable(summaryTable,
  rowVar = getAttribute(summaryTable, "rowVar"),
  rowVarLab = getAttribute(summaryTable, "rowVarLab", default =
  getLabelVar(rowVar, labelVars = labelVars)), rowVarInSepCol = NULL,
  rowVarFormat = NULL, rowVarTotalInclude = getAttribute(summaryTable,
  "rowVarTotalInclude"), rowVarTotalInSepRow = NULL,
  rowTotalLab = NULL, rowAutoMerge = TRUE,
  colVar = getAttribute(summaryTable, "colVar"),
  colTotalLab = getAttribute(summaryTable, "colTotalLab", default =
  "Total"), colHeaderTotalInclude = TRUE, labelVars = NULL,
  statsVar = getAttribute(summaryTable, "statsVar"),
  statsLayout = c("row", "col", "rowInSepCol"),
  statsValueLab = "StatisticValue", emptyValue = NULL,
  vline = c("none", "auto"))
}
\arguments{
\item{summaryTable}{Summary table, created with the \code{\link{computeSummaryStatisticsTable}} function.}

\item{rowVar}{Character vector with variable(s) used for the rows.
If multiple variables are specified, the variables should be sorted in hierarchical order.
The variables are included in rows, excepted if specified in \code{rowVarInSepCol}.}

\item{rowVarLab}{Label for the \code{rowVar} variable(s).}

\item{rowVarInSepCol}{Variable(s) of \code{rowVar} which should be 
included in separated column in the table, NULL by default. 
This is only available if \code{rowVar} if not specified.
Note that the row total (if \code{rowTotalInclude} is TRUE) is computed 
separately by this variable.}

\item{rowVarFormat}{Named list with special formatting for the \code{rowVar}.
Currently, only possibility is to set the variable elements in bold, with:
list(var1 = "bold").}

\item{rowVarTotalInclude}{Character vector with \code{rowVar}
for which to include the total for each group.}

\item{rowVarTotalInSepRow}{Character vector with \code{rowVarTotalInclude}
(not in \code{rowVarInSepCol}) for which the total should be included in a separated row labelled 'Total'.
Otherwise (by default) the total is included in the header row of each category.}

\item{rowTotalLab}{label for the row with total}

\item{rowAutoMerge}{Logical, if TRUE (by default) automatically merging of rows,
e.g. in case there is only one sub-category (e.g. categorical variable with only one group)
or only one statistic per category}

\item{colVar}{Character vector with variable(s) used for the columns.
If multiple variables are specified, the variables should be sorted in hierarchical order,
and are included in multi-columns layout.}

\item{colTotalLab}{String, label for the total column included 
if \code{colTotalInclude} is TRUE, 'Total' by default.}

\item{colHeaderTotalInclude}{Logical, if TRUE include the total of number of patients
(\code{'statN'}) in the header.}

\item{labelVars}{Named string with variable labels (names are the variable code).}

\item{statsVar}{Character vector with columns of \code{summaryTable} with
statistic variables.}

\item{statsLayout}{String with layout for the statistics names 
(in case more than one statistic is included), among:
\itemize{
\item{row: }{Statistics are included in rows in the first column of the table}
\item{'col': }{Statistics are included in columns (last row of the header).
This option is not compatible with categorical variable(s).}
\item{'rowInSepCol': }{Statistics are included in rows, but in a separated column than
the \code{rowVar} variable(s)}
}}

\item{statsValueLab}{String with label for the statistic value, 
in case no \code{colVar} is specified: 'StatisticValue' by default.}

\item{emptyValue}{Value used to fill the table for missing values.
See the \code{fill} parameter of the \code{\link[reshape2]{dcast}} function.}

\item{vline}{String mentioning how vertical lines should be included, either: 
\itemize{
\item{'none' (default): }{no vertical lines included}
\item{'auto': }{vertical lines included between sub-groups}
}}
}
\value{
summaryTable reformatted in long format, with extra attributes:
\itemize{
\item{'header': }{data.frame with header for each column}
\item{'padParams': }{list of list of parameters to be passed to the 
\code{\link[flextable]{padding}} function}
\item{'rowVar': }{column of output with row variable}
\item{'rowVarInSepCol': }{column(s) of output with row variable in separated column(s)}
\item{'vlineParams' and 'hlineParams': }{
list of list with correspondingly parameters for
vertical and horizontal lines
}
\item{'vline': }{\code{vline} parameter}
\item{'formatParams': }{list of list with special formatting for the table,
currently only used if \code{rowVarFormat} if specified.}
}
If \code{summaryTable} is a list of summary tables,
returns a list of corresponding summary tables in long format.
}
\description{
Format summary statistics table for export
}
\author{
Laure Cougnaud
}
