% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSummaryStatisticsTableToFlextable.R
\name{getGLPGFlextable}
\alias{getGLPGFlextable}
\title{Format a flextable to fulfill GLPG style.}
\usage{
getGLPGFlextable(data, ft = NULL, border = TRUE,
  fontname = switch(style, report = "Times", presentation = "Tahoma"),
  fontsize = switch(style, report = 8, presentation = 10),
  landscape = (style == "presentation"), style = "report",
  margin = 1, adjustWidth = TRUE, colorTable = getColorTable(style =
  style), align = TRUE, title = NULL, pageDim = NULL)
}
\arguments{
\item{data}{data.frame with data used in table.}

\item{ft}{Corresponding \code{\link[flextable]{flextable}}.}

\item{border}{Logical, if TRUE add a border.}

\item{fontname}{String with font name, 'Times' by default.}

\item{fontsize}{Integer with font size, 8 by default.}

\item{landscape}{Logical, if TRUE (by defaut) the table is presented in landscape
format.}

\item{style}{string with table style, either 'report' or 'presentation'}

\item{margin}{Margin in the document in inches.}

\item{adjustWidth}{Logical, if TRUE adjust column widths.}

\item{colorTable}{Named character vector with color for the table,
see output of \code{\link{getColorTable}} for required elements.}

\item{align}{Logical, if TRUE (by default), default alignment is set.}

\item{title}{Character vector with title(s) for the table.
Set to NULL (by default) if no title should be included.
Only available if \code{ft} is not specified.}

\item{pageDim}{Numeric vector of length 2 with page width and height in inches
in portrait format.}
}
\value{
\code{\link[flextable]{flextable}} with GLPG style.
}
\description{
Format a flextable to fulfill GLPG style.
}
\author{
Laure Cougnaud
}
