% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimals.R
\name{getMaxNDecimals}
\alias{getMaxNDecimals}
\title{Get maximum number of decimals in a variable,
based on pre-defined rule and/or data (see the function:
 \code{\link{getNDecimals}})}
\usage{
getMaxNDecimals(x, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{...}{Any parameters for the \code{\link{getNDecimals}} function.}
}
\value{
Integer with maximum number of decimals in a character vector.
}
\description{
Get maximum number of decimals in a variable,
based on pre-defined rule and/or data (see the function:
 \code{\link{getNDecimals}})
}
\author{
Laure Cougnaud
}
