% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimals.R
\name{getMaxNDecimalsData}
\alias{getMaxNDecimalsData}
\title{Get maximum number of decimals in a variable based on the data
(\code{\link{getNDecimalsData}})}
\usage{
getMaxNDecimalsData(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Integer with maximum number of decimals in a character vector.
}
\description{
Get maximum number of decimals in a variable based on the data
(\code{\link{getNDecimalsData}})
}
\author{
Laure Cougnaud
}
