% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimals.R
\name{getNDecimals}
\alias{getNDecimals}
\title{Get number of decimals for a specific vector,
either from specific implemented rule (see \code{\link{getNDecimalsRule}}),
or from the dataset itself (see \code{\link{getNDecimalsData}}),
or both: in this case the minimum of both criterias is used.}
\usage{
getNDecimals(x, useRule = TRUE, rule = "1", useData = TRUE)
}
\arguments{
\item{useRule}{Logical (TRUE by default), should the \code{rule} be applied?}

\item{useData}{Logical (TRUE by default), should the number of decimals
be extracted based on \code{\link{getNDecimalsData}}?}
}
\description{
Get number of decimals for a specific vector,
either from specific implemented rule (see \code{\link{getNDecimalsRule}}),
or from the dataset itself (see \code{\link{getNDecimalsData}}),
or both: in this case the minimum of both criterias is used.
}
\author{
Laure Cougnaud
}
