% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getNDecimals}
\alias{getNDecimals}
\title{Get number of decimals.}
\usage{
getNDecimals(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector of same length than \code{x}
with the number of decimals
}
\description{
Get number of decimals.
}
\author{
Laure Cougnaud
}
