% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimals.R
\name{getNDecimals}
\alias{getNDecimals}
\title{Get number of decimals for a specific vector.}
\usage{
getNDecimals(x, useRule = TRUE, rule = "1", useData = TRUE)
}
\arguments{
\item{x}{Numeric vector.}

\item{useRule}{Logical (TRUE by default), should the \code{rule} be applied?}

\item{rule}{Character vector with rule to use to derive 
the number of parameters.
Currently only: '1' is implemented.
\itemize{
\item{'1': }{Standard rule for the number of 
decimals for individual values for a 
continuous variable based on the 
\emph{Mock Standard SAP (version 1.0)} document.
\itemize{
\item{value < 1 ('very small values' in SAP): }{3}
\item{value < 10: }{2}
\item{value in [10, 1000[: }{1}
\item{value >= 1000: }{0}
}
}
}}

\item{useData}{Logical (TRUE by default), should the number of decimals
be extracted based on the input data \code{x}?}
}
\value{
Numeric vector of same length than \code{x}
with the number of decimals.
}
\description{
The number of decimals is extracted from specific implemented rule
(see \code{\link{getNDecimalsRule}}),
from the data itself (see \code{\link{getNDecimalsData}}),
or both: in this case the minimum of both criterias is used.
}
\author{
Laure Cougnaud
}
