% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimals.R
\name{getNDecimalsData}
\alias{getNDecimalsData}
\title{Get number of decimals based on the data in a numeric vector.
Note: NA is returned if the element is missing (NA).}
\usage{
getNDecimalsData(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector of same length than \code{x}
with the number of decimals.
}
\description{
Get number of decimals based on the data in a numeric vector.
Note: NA is returned if the element is missing (NA).
}
\examples{
x <- c(0.99, 5.679, 50.45, 1450)
getNDecimalsData(x)
}
\seealso{
Other decimals: 
\code{\link{formatPercentage}()},
\code{\link{getMaxNDecimalsData}()},
\code{\link{getMaxNDecimals}()},
\code{\link{getNDecimals}()}
}
\author{
Laure Cougnaud
}
\concept{decimals}
