% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_stats.R
\name{getStats}
\alias{getStats}
\title{Create custom statistic sets to be passed to the \code{stats} parameter
of the \code{\link{computeSummaryStatisticsTable}} function.}
\usage{
getStats(type = c("summary", "count", "n (\%)"))
}
\arguments{
\item{type}{String with type of statistics:
\itemize{
\item{'summary': }{all statistics for 'summaryTable' (\code{type} parameter)}
\item{'count': }{all statistics for 'countTable' (\code{type} parameter)}
\item{'n (\%)':}{number of subjects with 0 digit (percentage with 1 digits)}
}}
}
\value{
(Optionally named) list of expression or call object of summary statistics of interest
}
\description{
Create custom statistic sets to be passed to the \code{stats} parameter
of the \code{\link{computeSummaryStatisticsTable}} function.
}
\author{
Laure Cougnaud
}
