% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_stats.R
\name{getStatsData}
\alias{getStatsData}
\title{Get default set of statistics for variables of interest and specific dataset}
\usage{
getStatsData(data, var = NULL, type = c("default", "all"), extra = NULL, ...)
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{Character vector with variables of interest}

\item{type}{String with type of statistics to extract:
\itemize{
\item{'default': }{default sets of statistics, 
see types: 'summary-default' and 'count-default' in \code{\link{getStats}}}
\item{'all': }{all computed statistics, see types: 'summary' and 'count' in \code{\link{getStats}}}
}}

\item{extra}{List with extra statistics to include, or function to apply on each
\code{var} (e.g. depending on the class of \code{var}) to get such list.}

\item{...}{parameters passed to the \code{\link{getStats}} function}
}
\value{
List with statistics to compute, named by \code{var}
}
\description{
This set of statistics is passed directly to the \code{stats} parameter
of the \code{\link{computeSummaryStatisticsTable}} function.
}
\examples{
library(glpgUtilityFct)
data(ADaMDataPelican)
getStatsData(data = ADaMDataPelican$ADSL, var = "WEIGHTBL")
}
\author{
Laure Cougnaud
}
