% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryStatisticsTable.R
\name{getSummaryStatisticsTable}
\alias{getSummaryStatisticsTable}
\title{Get summary statistics for a specific dataset and variables of interest}
\usage{
getSummaryStatisticsTable(data, var = "AVAL", colVar = NULL,
  rowVar = NULL, subjectVar = "USUBJID", totalLabel = "Total")
}
\arguments{
\item{data}{data.frame with data}

\item{var}{string, variable of \code{data} with variable to use,
'AVAL' by default}

\item{colVar}{variable(s) of \code{data} used 
for grouping in column. The total 
for each subgroup across \code{rowVar} is computed.}

\item{rowVar}{variable(s) of \code{data} used for
grouping in row.}

\item{subjectVar}{string, variable of \code{data} with subject ID,
'USUBJID' by default}

\item{totalLabel}{string label for the column with the total,
in case \code{totalInclude} is TRUE,
'total' by default.}

\item{totalInclude}{logical, if TRUE (by default) and \code{by} is specified,
include also the summary statistics for the entire dataset}
}
\value{
data.frame with summary statistics in columns:
\itemize{
\item{'N': }{number of subjects}
\item{'Mean': }{mean of \code{var}}
\item{'SD': }{standard deviation of \code{var}}
\item{'SE': }{standard error of \code{var}}
\item{'Median': }{median of \code{var}}
\item{'Min': }{minimum of \code{var}}
\item{'Max': }{maximum of \code{var}}
}
}
\description{
Get summary statistics for a specific dataset and variables of interest
}
\author{
Laure Cougnaud
}
