% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryStatisticsTable.R
\name{getSummaryStatisticsTable}
\alias{getSummaryStatisticsTable}
\title{Get summary statistics table}
\usage{
getSummaryStatisticsTable(
  data,
  var = NULL,
  varFlag = NULL,
  varLab = NULL,
  varLabInclude = length(var) > 1,
  varInclude0 = FALSE,
  varIgnore = NULL,
  varGeneralLab = "Variable",
  varSubgroupLab = "Variable group",
  varIncludeTotal = FALSE,
  varTotalInclude = FALSE,
  varTotalInSepRow = FALSE,
  rowVar = NULL,
  rowVarLab = NULL,
  rowVarDataLevels = NULL,
  rowOrder = "auto",
  rowOrderTotalFilterFct = NULL,
  rowOrderCatLast = NULL,
  rowVarInSepCol = NULL,
  rowVarFormat = NULL,
  rowVarTotalInclude = NULL,
  rowVarTotalByVar = NULL,
  rowVarTotalInSepRow = NULL,
  rowTotalLab = NULL,
  rowInclude0 = FALSE,
  rowAutoMerge = TRUE,
  emptyValue = "-",
  rowVarTotalPerc = NULL,
  colVar = NULL,
  colVarTotal = colVar,
  colVarTotalPerc = colVarTotal,
  colInclude0 = FALSE,
  colVarDataLevels = NULL,
  colTotalInclude = FALSE,
  colTotalLab = "Total",
  stats = NULL,
  statsExtra = NULL,
  statsVarBy = NULL,
  statsPerc = c("statN", "statm"),
  statsGeneralLab = "Statistic",
  statsValueLab = "StatisticValue",
  statsLabInclude = NULL,
  subjectVar = "USUBJID",
  filterFct = NULL,
  dataTotal = NULL,
  dataTotalPerc = dataTotal,
  dataTotalRow = NULL,
  dataTotalCol = NULL,
  type = "auto",
  labelVars = NULL,
  landscape = (style == "presentation"),
  margin = 1,
  rowPadBase = 14.4,
  title = NULL,
  footer = NULL,
  file = NULL,
  outputType = "flextable",
  statsLayout = ifelse("DT" \%in\% outputType, "col", "row"),
  style = "report",
  colorTable = getColorTable(style = style),
  byVar = NULL,
  byVarLab = NULL,
  colHeaderTotalInclude = TRUE,
  fontsize = switch(style, report = 8, presentation = 10),
  fontname = switch(style, report = "Times", presentation = "Tahoma"),
  vline = "none",
  hline = "auto",
  pageDim = NULL,
  expandVar = NULL,
  noEscapeVar = NULL,
  barVar = NULL,
  ...
)
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{Character vector, variable(s) of \code{data} to compute statistics on.
Missing values, if present, are filtered.
If NULL (by default), counts of the \code{rowVar} are returned.
To also return counts of the \code{rowVar} in case other \code{var}
are specified, you can include: 'all' in the \code{var}.}

\item{varFlag}{Character vector, subset of \code{var} with variable(s) 
of type 'flag' (with 'Y', 'N' or '' for empty/non specified value).
Only the counts for records flagged (with 'Y') are retained.}

\item{varLab}{Label for the \code{var} variable.}

\item{varLabInclude}{Logical, if TRUE
the name of the summary statistic variable(s) (\code{var})
are included in the table.
This is automatically set to TRUE if more than one variable(s) 
and is specified, and FALSE if only one variable is specified.}

\item{varInclude0}{Logical, should rows with no counts 
for the count \code{var} or \code{varFlag} variable(s)
be included in the table?
Either:
\itemize{
\item{logical of length 1, if TRUE (FALSE by default) 
rows with no count are included for all \code{var}
}
\item{a character vector containing categorical \code{var} 
for which zero counts rows should be included}
}}

\item{varIgnore}{Vector with elements to ignore in the \code{var} variable(s).
The \code{data} records with such elements in \code{var} are \strong{filtered} from the data 
at the start of the workflow.}

\item{varGeneralLab}{String with general label for variable specified in \code{var}.
In case of multiple variable in \code{var}, this will be included in the table header
(see 'rowVarLab' attribute of the output).}

\item{varSubgroupLab}{String with general label for sub-group of
categorical variable(s) for count table, 'Variable group' by default.
This will be included in the final table header (see 'rowVarLab' attribute of the output).}

\item{varIncludeTotal}{This argument is deprecated, please use: 'varTotalInclude' instead.}

\item{varTotalInclude}{Should the total across all categories of \code{var} 
be included for the count table?
Only used for categorical variables (and \code{var} not 'all').
Either:
\itemize{
\item{logical of length 1, if TRUE (FALSE by default) include the total for all categorical \code{var}}
\item{a character vector containing categorical \code{var} for which the total should be included}
}}

\item{varTotalInSepRow}{Logical, should the total per variable be included in
a separated row (by default) or in the row containing the header of the variable?}

\item{rowVar}{Variable(s) of \code{data} used for
grouping in row in the final table.}

\item{rowVarLab}{Label for the \code{rowVar} variable(s).}

\item{rowVarDataLevels}{Data.frame with unique combinations of \code{rowVar}
to be included in columns.
Each column should correspond to \code{colVar} and as factor
if the elements should be ordered in the final table.}

\item{rowOrder}{Specify how the rows should be ordered in the final table, either a:
\itemize{
\item{String among:}{
\itemize{
\item{'auto' (by default): }{if the variable is a factor, keep its order, otherwise order alphabetically}
\item{'alphabetical': }{order alphabetically}
\item{'total': }{order rows in decreasing order of the total number of subjects
across all columns for this specific category.}
}}
\item{Function with input the summary table and output the ordered elements of the \code{rowVar}}
}
To specify different ordering methods for different \code{rowVar}, specify a list
of such elements, named with the \code{rowVar} variable.
For the table output of \code{\link{computeSummaryStatisticsTable}} (long format),
this order is also reflected in the \strong{\code{levels}} of the row factor variable.}

\item{rowOrderTotalFilterFct}{Function used to filter the data used to order the rows
based on total counts (in case \code{rowOrder} is 'total'),
To order rows based on one specific column category,
e.g. to order based on the counts in the treatment column:
function(x) subset(x, TRTP == "treatmentX")}

\item{rowOrderCatLast}{String with category to be printed in the last 
row of each \code{rowVar} (if any, set to NULL if none).}

\item{rowVarFormat}{Parameters specific to the 'flextable' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToFlextable}}.}

\item{rowVarTotalInclude}{Character vector with \code{rowVar}
for which the total should be reported.\cr
If the higher row variable is specified, the total across all rows
is reported.}

\item{rowVarTotalByVar}{Character vector with a row variable
used to categorize the row total.\cr
Note that this is only used if row total(s) is/are requested via \code{rowVarTotalInclude},
and this variable should also be included in \code{rowVar}.
This can be specified also for a specific row variable if the vector is named.\cr
For example: \code{c(ADECOD = "AESEV")} to compute total by severity 
for row adverse event term in a typical adverse event count table 
(by System Organ Class and Adverse Event Term).}

\item{rowVarTotalInSepRow}{Character vector with \code{rowVarTotalInclude}
(not in \code{rowVarInSepCol}) for which the total should be included in a separated row labelled 'Total'.
Otherwise (by default) the total is included in the header row of each category.}

\item{rowTotalLab}{Parameters specific to the 'flextable' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToFlextable}}.}

\item{rowInclude0}{Logical, if TRUE (FALSE by default),
include rows with no records, based on all combinations 
of the \code{rowVar} (assuming nested variable(s)).}

\item{rowAutoMerge}{Parameters specific to the 'flextable' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToFlextable}}.}

\item{emptyValue}{String with placeholder used to fill the table for missing values, '-' by default.
This value is typically used e.g. if not all statistics are computed for all specified
row/col/var variables.}

\item{rowVarTotalPerc}{Character vector with row variables by which the total
should be computed for the denominator for the percentage computation.
By default the total is only computed only by column (NULL by default).
If the total should be based on the total number of records per variable,
\code{rowVarTotalPerc} should be set to 'variable'.}

\item{colVar}{Character vector with variable(s) used for the columns.
If multiple variables are specified, the variables should be sorted in hierarchical order,
and are included in multi-columns layout.
Use: 'variable' to include the variables to summarize: \code{var}
 (if multiple) in different columns.}

\item{colVarTotal}{String with column(s) considered to compute the total by,
reported in the header of the table, by default same as \code{colVar}.
Use: 'variable' to compute total by \code{var} (if multiple).}

\item{colVarTotalPerc}{String with column(s) considered to compute the total by,
used as denominator for the percentage computation, by default same as \code{colVarTotal}.
Use: 'variable' to compute total by \code{var} (if multiple).}

\item{colInclude0}{Logical, if TRUE (FALSE by default),
include columns with no records, based on all combinations 
of the \code{columnVar} (assuming nested variable(s)).
If variable(s) are not nested, possible combinations
can be specified via \code{colVarDataLevels}.}

\item{colVarDataLevels}{Data.frame with unique combinations of \code{colVar}
to be included in columns.
Each column should correspond to \code{colVar} and as factor
if the elements should be ordered in the final table.}

\item{colTotalInclude}{Logical, if TRUE (FALSE by default) include the summary 
statistics across columns in a separated column.}

\item{colTotalLab}{String, label for the total column included 
if \code{colTotalInclude} is TRUE, 'Total' by default.}

\item{stats}{(optional) Statistic(s) of interest to compute, either:
\itemize{
\item{string with a default set of statistics, 
see section 'Formatted statistics' in
\code{\link[=inTextSummaryTable-stats]{in-text table statistics}}.}
\item{Named list of expressions or call objects (see \code{\link{is.language}}) 
of base summary statistics of interest, see section:
'Base statistics' in
\code{\link[=inTextSummaryTable-stats]{in-text table statistics}}.\cr
The names are reported in the header.\cr
If \code{stats} if of length 1, the name of the summary statistic is not included
in the table.\cr
The statistics can be specified separately:
\itemize{
\item{for each \code{var} (if multiple), 
by naming each element of the list:
\code{list(varName1 = list(...), varName2 = list())}
}
\item{and/or for each element in:
\code{statsVarBy}, by naming each sublist.}
}}
}}

\item{statsExtra}{(optional) Named list with functions for additional custom
statistics to be computed, only available for 'summaryTable',
e.g. list(statCVPerc = function(x) sd(x)/mean(x)*100) (or \code{\link{cv}}).
Each function has as parameter: either 'x': the variable or 'data': the entire dataset,
and return the corresponding summary statistic.}

\item{statsVarBy}{String with variable in \code{rowVar}/code{colVar}
which the statistics should be computed by.
In this case, \code{stats} (nested list or not) should be additionally nested
to specify the statistics for each element in \code{statsVarBy}.}

\item{statsPerc}{String with 'base statistical variable' used to compute the 
percentage, either: 
\itemize{
\item{'statN' (by default): }{the number of subjects}
\item{'statm': }{the number of records}
}}

\item{statsGeneralLab}{String with general label for statistics, 'Statistic' by default.
Only included if no \code{statsVar} if longer than 1.}

\item{statsValueLab}{String with label for the statistic value, 
'StatisticValue' by default.
This is only used if the statistics provided in \code{stats} are not named
and if no \code{colVar} is specified.}

\item{statsLabInclude}{Logical, if TRUE include the statistic label
in the table. By default only included if more than
one statistic variables are available in the table.}

\item{subjectVar}{String, variable of \code{data} with subject ID,
'USUBJID' by default.}

\item{filterFct}{(optional) Function taking as input
the summary table with computed statistics and returning a subset 
of the summary table.\cr
Note: The filtering function should also handle records with :
\itemize{
\item{total for the column header: \code{isTotal} set to TRUE,
and \code{colVar}/\code{rowVar} is NA.\cr
For example: \code{filterFct = function(data) subset(data, isTotal & myColVar == "group 1")}
}
\item{\code{rowVar}/\code{colVar} set to 'Total'/\code{colTotalLab} 
if \code{rowVarTotalInclude}/\code{colTotalInclude} is specified}
}}

\item{dataTotal}{Data.frame used to extract the Total number of subject
per column in column header ('N = [X]').
It should contain the variables specified by \code{colVarTotal}.
If not specified, the total number of subjects is extracted from the \code{data}.}

\item{dataTotalPerc}{Data.frame used to extract the total counts per column 
for the computation of the percentage.\cr
By default, \code{dataTotal} is used.\cr
It should contain the variables specified by \code{colVarTotalPerc}.}

\item{dataTotalRow}{Data.frame used to extract the total count across all
elements of the row
variable, list of such data.frame for each \code{rowVar} variable.\cr
If the dataset is specified by row variable, the list should be named with:
variable X if the total across elements of variable X should be included.
By default, \code{data} is used.}

\item{dataTotalCol}{Data.frame from which the total across columns is 
extracted (in case \code{colTotalInclude} is TRUE)
or list of such data.frame for each \code{rowVar} variable.\cr
If the dataset is specified by row variable, the list should be named with:
with:
\itemize{
\item{last row variable: }{for the dataset used in the total column for 
the most nested row variable}
\item{higher row variable (X+1): }{for the dataset used for the total column
and row total of X}
\item{'total': }{for the dataset used for the total column and general row total}
}
If only a subset of the variables is specified in this list, 
\code{data} is used for the remaining variable(s) (or 'total') if needed.\cr
This dataset (the one for 'total' if a list) is also used for:
\itemize{
\item{the header of the total column in case \code{dataTotal} is
not specified}
\item{the denominator of the percentages in the total column
in case \code{dataTotalPerc} is not specified}
}
By default, \code{data} is used.}

\item{type}{String with type of table: 
\itemize{
\item{'summaryTable': }{summary table with custom statistics}
\item{'countTable': }{count table}
\item{'auto' (by default): }{'summaryTable' of \code{var} is numeric,
'countTable' otherwise
}
}}

\item{labelVars}{Named character vector with variable labels (names are the variable code),
usually extracted from \code{data}.}

\item{landscape}{Parameters specific to the 'flextable' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToFlextable}}.}

\item{margin}{Parameters specific to the 'flextable' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToFlextable}}.}

\item{rowPadBase}{Parameters specific to the 'flextable' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToFlextable}}.}

\item{footer}{Parameters specific to the 'flextable' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToFlextable}}.}

\item{file}{(Optional) Name of the file the table should be exported to, 
either:
\itemize{
\item{string (of length 1). In this case, depending on the
file extension, the following is exported: }{
\itemize{
\item{'txt': summary table in long format ('data.frame-base' \code{outputType})}
\item{'docx': summary table in final format is exported ('flextable' \code{outputType})}
\item{'html': interactive summary table is exported ('DT' \code{outputType})}
}}
\item{named character vector in case of multiple exports.
The names should correspond to the options in \code{outputType}:}{
\itemize{
\item{for 'data.frame-base' and 'data.frame': }{filename with 'txt' extension}
\item{for 'flextable': }{filename with 'docx' extension}
\item{for 'DT': }{filename with 'html' extension}
}}
}
If NULL (by default), the summary table is not exported but only returned as output.
If \code{byVar} is specified, each table is exported to a separated
file with the suffix: 'file_[i].[ext]' with i the index of the file.}

\item{outputType}{String with output type:
\itemize{
\item{'flextable' (by default): }{\code{\link[flextable]{flextable}} object, with format for
CSR, compatible with Word/PowerPoint export}
\item{'DT': }{\code{\link[DT]{datatable}} interactive table,
compatible with html export}
\item{'data.frame': }{data.frame in wide format (with elements in 
\code{colVar} in different columns)}
\item{'data.frame-base'}{data.frame in long format (with elements in 
\code{colVar} in different rows), useful for QC}
}}

\item{statsLayout}{String with layout for the statistics names 
(in case more than one statistic is included), among:
\itemize{
\item{row (by default when \code{outputType} is: 'flextable'): }{
All statistics are included in different rows 
in the first column of the table
}
\item{'col' (by default when \code{outputType} is: 'DT'): }{
Statistics are included in separated columns (last row of the header).
This option is not compatible with categorical variable(s).
}
\item{'rowInSepCol': }{
Statistics are included in different rows, but in a separated column than
the \code{rowVar} variable(s)
}
}}

\item{style}{Parameters specific to the 'flextable' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToFlextable}}.}

\item{colorTable}{Parameters specific to the 'flextable' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToFlextable}}.}

\item{byVar}{Variable(s) of \code{data} for which separated table(s)
should be created.}

\item{byVarLab}{String with label for \code{byVar}, used to set the names
of the output list of table(s).}

\item{colHeaderTotalInclude}{Logical, if TRUE include the total of number of patients
(\code{'statN'}) in the header.}

\item{fontsize}{Parameters specific to the 'flextable' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToFlextable}}.}

\item{fontname}{Parameters specific to the 'flextable' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToFlextable}}.}

\item{vline}{Parameters specific to the 'flextable' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToFlextable}}.}

\item{hline}{Parameters specific to the 'flextable' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToFlextable}}.}

\item{pageDim}{Numeric vector of length 2 with page width and height
depending on \code{outputType}:
\itemize{
\item{'flextable': }{in inches in portrait format}
\item{'DT': }{number of rows in the table (currently only
the height is used (e.g. \code{c(NA, 4)})
}}}

\item{expandVar}{Parameters specific to the 'DT' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToDT}}.}

\item{noEscapeVar}{Parameters specific to the 'DT' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToDT}}.}

\item{barVar}{Parameters specific to the 'DT' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToDT}}.}

\item{...}{Parameters specific to the 'DT' \code{output}.\cr 
See correspondings parameter in \code{\link{exportSummaryStatisticsTableToDT}}.}
}
\value{
Depending on the \code{outputType}:
\itemize{
\item{'data.frame-base': }{input summary table in a long format with
all computed statistics}
\item{'data.frame': }{summary table in a wide format (
different columns for each \code{colVar}), with specified labels}
\item{'flextable' (by default): }{\code{\link[flextable]{flextable}}
 object with summary table}
\item{'DT': }{\code{\link[DT]{datatable}} object with summary table}
}
If multiple \code{outputType} are specified, a list of those objects, named
by \code{outputType}.
If \code{byVar} is specified, each object consists of a list of tables,
one for each element in \code{byVar}.
}
\description{
Get summary statistics table
}
\author{
Laure Cougnaud
}
