% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryStatisticsTable.R
\name{getSummaryStatisticsTable}
\alias{getSummaryStatisticsTable}
\title{Get summary statistics table}
\usage{
getSummaryStatisticsTable(data, var = NULL, varLab = getLabelVar(var,
  data = data, labelVars = labelVars), varIgnore = NULL,
  varGeneralLab = "Variable", varSubgroupLab = NULL,
  varIncludeTotal = FALSE, rowVar = NULL,
  rowVarLab = getLabelVar(rowVar, labelVars = labelVars),
  rowVarDataLevels = NULL, rowOrder = "auto",
  rowOrderTotalFilterFct = NULL, rowVarInSepCol = NULL,
  rowVarTotalInclude = NULL, rowVarTotalByVar = NULL,
  rowVarTotalInSepRow = NULL, rowTotalLab = NULL,
  rowInclude0 = FALSE, rowAutoMerge = TRUE, emptyValue = NULL,
  rowVarTotalPerc = NULL, colVar = NULL, colVarTotal = colVar,
  colVarTotalPerc = colVarTotal, colInclude0 = FALSE,
  colVarDataLevels = NULL, colTotalInclude = FALSE,
  colTotalLab = "Total", stats = NULL, statsExtra = NULL,
  statsVarBy = NULL, statsGeneralLab = "Statistic",
  statsValueLab = "StatisticValue", subjectVar = "USUBJID",
  filterFct = NULL, dataTotal = NULL, dataTotalPerc = dataTotal,
  type = "auto", labelVars = NULL, landscape = (style ==
  "presentation"), margin = 1, rowPadBase = 14.4, title = NULL,
  footer = NULL, file = NULL, outputType = c("flextable",
  "data.frame"), statsLayout = c("row", "col", "rowInSepCol"),
  style = "report", colorTable = getColorTable(style = style),
  byVar = NULL, byVarLab = getLabelVar(byVar, data = data, labelVars =
  labelVars), colHeaderTotalInclude = TRUE, fontsize = switch(style,
  report = 8, presentation = 10), fontname = switch(style, report =
  "Times", presentation = "Tahoma"), vline = "none", pageDim = NULL)
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{Character vector with variable, used if 'variable' is specified
within \code{rowVarTotal}.}

\item{varLab}{Label for the \code{var} variable.}

\item{varIgnore}{Vector with elements to ignore in the \code{var} variable}

\item{varGeneralLab}{String with general label for variable specified in \code{var}.
In case of multiple variable in \code{var}, this will be included in the table header
(see 'rowVarLab' attribute of the output).}

\item{varSubgroupLab}{String with general label for sub-group, in case
\code{var} is specified for a count table.
This will be included in the table header (see 'rowVarLab' attribute of the output).
Empty by default.}

\item{varIncludeTotal}{Logical, if TRUE (FALSE by default) the total across 
all categories of \code{var} is included.
Only considered if \code{type} is 'countTable'.}

\item{rowVar}{Character vector with variable(s) used for the rows.
If multiple variables are specified, the variables should be sorted in hierarchical order.
The variables are included in rows, excepted if specified in \code{rowVarInSepCol}.}

\item{rowVarLab}{Label for the \code{rowVar} variable(s).}

\item{rowVarDataLevels}{Data.frame with levels to consider for the \code{rowVar} variable(s)
(in case of non nested row variable(s)).}

\item{rowOrder}{String or function of named list with method used to order the rows,
see the \code{method} parameter of \code{\link{convertVarToFactorWithOrder}}.
'total' is only available if the total is available for each \code{rowVar} (\code{rowSubtotalInclude} set to TRUE)
If a string, the same method is used for all \code{rowVar},
otherwise the list is named with the \code{rowVar} variable, to
specify a different ordering method for each variable.}

\item{rowOrderTotalFilterFct}{Function used to filter the data used to order the rows
based on total counts (in case \code{rowOrder} is 'total').}

\item{rowVarInSepCol}{Variable(s) of \code{rowVar} which should be 
included in separated column in the table, NULL by default. 
This is only available if \code{rowVar} if not specified.
Note that the row total (if \code{rowTotalInclude} is TRUE) is computed 
separately by this variable.}

\item{rowVarTotalInclude}{Character vector with \code{rowVar}
for which to include the total for each group.}

\item{rowVarTotalByVar}{Character vector with extra \code{codeVar}
for which to compute the row total by.
Can be specified for each \code{rowVarTotalInclude} if named by the corresponding variable.}

\item{rowVarTotalInSepRow}{Character vector with \code{rowVarTotalInclude}
(not in \code{rowVarInSepCol}) for which the total should be included in a separated row labelled 'Total'.
Otherwise (by default) the total is included in the header row of each category.}

\item{rowTotalLab}{label for the row with total}

\item{rowInclude0}{Logical, if TRUE (FALSE by default),
include rows with no records, based on all combinations 
of the \code{rowVar} (assuming nested variable(s)).}

\item{rowAutoMerge}{Logical, if TRUE (by default) automatically merging of rows,
e.g. in case there is only one sub-category (e.g. categorical variable with only one group)
or only one statistic per category}

\item{emptyValue}{Value used to fill the table for missing values.
See the \code{fill} parameter of the \code{\link[reshape2]{dcast}} function.}

\item{rowVarTotalPerc}{Character vector with row(s) considered to compute the total,
used as denominator for the percentage computation.
By default the total is only computed by column (NULL by default).
If the total should be based on the total number of records per variable,
\code{rowVarTotalPerc} should be set to 'variable'.}

\item{colVar}{Character vector with variable(s) used for the columns.
If multiple variables are specified, the variables should be sorted in hierarchical order,
and are included in multi-columns layout.}

\item{colVarTotal}{String with column(s) considered to compute the total by,
reported in the header of the table, by default same as \code{colVar}.}

\item{colVarTotalPerc}{String with column(s) considered to compute the total by,
used as denominator for the percentage computation, by default same as \code{colVarTotal}.}

\item{colInclude0}{Logical, if TRUE (FALSE by default),
include columns with no records, based on all combinations 
of the \code{columnVar} (assuming nested variable(s)).
If variable(s) are not nested, possible combinations
can be specified via \code{colVarDataLevels}.}

\item{colVarDataLevels}{Data.frame with levels to consider for the \code{colVar} variable(s)
(in case of non nested column variable(s)).}

\item{colTotalInclude}{Logical, if TRUE (FALSE by default) include the summary 
statistics across columns in a separated column.}

\item{colTotalLab}{String, label for the total column included 
if \code{colTotalInclude} is TRUE, 'Total' by default.}

\item{stats}{(Optionally) named list of expression or call object of summary statistics of interest.
The names are reported in the header.
The following variables are recognized, if the table is a: 
\itemize{
\item{'summaryTable': }{
\itemize{
\item{'statN': }{number of subjects}
\item{'statMean': }{mean of \code{var}}
\item{'statSD': }{standard deviation of \code{var}}
\item{'statSE': }{standard error of \code{var}}
\item{'statMedian': }{median of \code{var}}
\item{'statMin': }{minimum of \code{var}}
\item{'statMax': }{maximum of \code{var}}
\item{'statPerc': }{percentage of subjects}
\item{'statm': }{number of records}
}
}
\item{'countTable': }{
\itemize{
\item{'statN': }{number of subjects}
\item{'statPercN': }{percentage of subjects}
\item{'statm': }{number of records}
}
}
}
If \code{stats} if of length 1, the name of the summary statistic is not included
in the table.
The statistics can be specified for each \code{var} (if multiple), 
by naming each element of the list:
list(varName1 = list(...), varName2 = list()) and/or for each element in:
\code{statsVarBy}, by naming each sublist.}

\item{statsExtra}{(optional) Named list with functions for additional custom
statistics to be computed, only available for 'summaryTable',
e.g. list(statCVPerc = function(x) sd(x)/mean(x)*100) (or \code{\link{cv}}).
Each function has as parameter: either 'x': the variable or 'data': the entire dataset,
and return the corresponding summary statistic.}

\item{statsVarBy}{String with variable in \code{rowVar}/code{colVar}
which the statistics should be computed by.
In this case, \code{stats} (nested list or not) should be additionally nested
to specify the statistics for each element in \code{statsVarBy}.}

\item{statsGeneralLab}{String with general label for statistics, 'Statistic' by default.
Only included if no \code{statsVar} if longer than 1.}

\item{statsValueLab}{String with label for the statistic value, 
in case no \code{colVar} is specified: 'StatisticValue' by default.}

\item{subjectVar}{String, variable of \code{data} with subject ID,
'USUBJID' by default.}

\item{filterFct}{(optional) Function based on computed statistics of
\code{rowVar}/code{colVar} which returns a subset of the summary table 
(after statistics computation).
Note: The filtering function should also handle records with :
\itemize{
\item{\code{rowVar}/\code{codeVar} set to 'Total'/\code{colTotalLab} 
if \code{rowVarTotalInclude}/\code{colTotalInclude} is TRUE}
\item{total for the column header: \code{isTotal} set to TRUE,
and \code{colVar}/\code{rowVar} is NA}
}
Note: the total count per category of the row variables is included
in a row with this variable set to 'Total'.}

\item{dataTotal}{Data.frame used to extract the Total count, indicated
in 'N' in column header.
It should contain the variables specified by \code{colVarTotal}.}

\item{dataTotalPerc}{Data.frame used to extract the Total count, 
used for the computation of the percentage: 'statPercN' parameter.
It should contain the variables specified by \code{colVarTotalPerc}.}

\item{type}{String with type of table: 
\itemize{
\item{'summaryTable': }{summary table with custom statistics}
\item{'countTable': }{count table}
\item{'auto' (by default): }{'summaryTable' of \code{var} is numeric,
'countTable' otherwise
}
}}

\item{labelVars}{Named string with variable labels (names are the variable code).}

\item{landscape}{Logical, if TRUE (by defaut) the table is presented in landscape
format.}

\item{margin}{Margin in the document in inches, currently only used to specify the
width of the table: [width page extracted from \code{dimPage} - 2* margin]}

\item{rowPadBase}{Base padding for row (in points), 14.4 by default (corresponds to 0.2 inches)}

\item{title}{Character vector with title(s) for the table.
Set to NULL (by default) if no title should be included.}

\item{footer}{Character vector with footer(s) for the table.
Set to NULL (by default) of no footer should be included.}

\item{file}{String with path of the file where the table should be exported.
If NULL, the summary table is not exported but only returned as output.}

\item{outputType}{String with output type, 'data.frame' or 'flextable'.}

\item{statsLayout}{String with layout for the statistics names 
(in case more than one statistic is included), among:
\itemize{
\item{row: }{Statistics are included in rows in the first column of the table}
\item{'col': }{Statistics are included in columns (last row of the header).
This option is not compatible with categorical variable(s).}
\item{'rowInSepCol': }{Statistics are included in rows, but in a separated column than
the \code{rowVar} variable(s)}
}}

\item{style}{string with table style in case \code{outputType} is 'flextable',
 either 'report' or 'presentation'.
This parameter affects the fontsize, font family, color of the text and background, 
and table dimensions of the table.}

\item{byVar}{Variable(s) of \code{data} for which separated table(s)
should be created.}

\item{byVarLab}{String with label for \code{byVar}, used to set the names
of the output list of table(s).}

\item{colHeaderTotalInclude}{Logical, if TRUE include the total of number of patients
(\code{'statN'}) in the header.}

\item{fontsize}{Integer with font size, by default:
8 if \code{style} is 'report' and 10 if \code{style} is 'presentation'.}

\item{fontname}{String with font name, by default:
'Times' if \code{style} is 'report' and 'Tahoma' if \code{style} is 'presentation'.}

\item{vline}{String mentioning how vertical lines should be included, either: 
\itemize{
\item{'none' (default): }{no vertical lines included}
\item{'auto': }{vertical lines included between sub-groups}
}}

\item{pageDim}{Numeric vector of length 2 with page width and height in inches
in portrait format.}
}
\value{
Depending on the \code{outputType}:
\itemize{
\item{'flextable': }{\code{\link[flextable]{flextable}} object with summary table}
\item{'data.frame': }{data.frame with summary table}
}
If \code{summaryTable} is a list of summary tables,
returns a list of corresponding summary tables in long format.
}
\description{
Get summary statistics table
}
\author{
Laure Cougnaud
}
