% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryStatisticsTable.R
\name{getSummaryStatisticsTable}
\alias{getSummaryStatisticsTable}
\title{Get summary statistics table}
\usage{
getSummaryStatisticsTable(data, var = NULL, varLab = getLabelVar(var,
  data = data, labelVars = labelVars), varIgnore = NULL,
  varLabGeneral = "Variable", varLabSubgroup = NULL,
  varIncludeTotal = FALSE, rowVar = NULL,
  rowVarLab = getLabelVar(rowVar, labelVars = labelVars),
  rowVarDataLevels = NULL, rowOrder = "auto",
  rowOrderTotalFilterFct = NULL, rowVarInSepCol = NULL,
  rowTotalInclude = FALSE, rowTotalLab = NULL,
  rowSubtotalInclude = FALSE, rowInclude0 = FALSE, emptyValue = NULL,
  colVar = NULL, colVarTotal = colVar, colVarTotalPerc = colVarTotal,
  colInclude0 = FALSE, colVarDataLevels = NULL,
  colTotalInclude = FALSE, colTotalLab = "Total",
  subjectVar = "USUBJID", stats = NULL, statsExtra = NULL,
  statsValueLab = "StatisticValue", filterFct = NULL,
  dataTotal = NULL, dataTotalPerc = dataTotal, type = "auto",
  labelVars = NULL, landscape = (style == "presentation"),
  margin = 1, rowPadBase = 4, title = NULL, footer = NULL,
  file = NULL, outputType = c("flextable", "data.frame"),
  statsLayout = c("row", "col", "rowInSepCol"), style = "report",
  byVar = NULL, byVarLab = getLabelVar(byVar, data = data, labelVars =
  labelVars), colHeaderTotalInclude = TRUE, fontsize = switch(style,
  report = 8, presentation = 10))
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{String, variable of \code{data} with variable to compute statistics on.
Missing values, if present, are filtered.}

\item{varLab}{Label for the \code{var} variable.}

\item{varIgnore}{Vector with elements to ignore in the \code{var} variable}

\item{varLabGeneral}{String with general label for variable specified in \code{var}.
In case of multiple variable in \code{var}, this will be included in the table header
(see 'rowVarLab' attribute of the output).}

\item{varLabSubgroup}{String with general label for sub-group, in case
\code{var} is specified for a count table.
This will be included in the table header (see 'rowVarLab' attribute of the output).
Empty by default.}

\item{varIncludeTotal}{Logical, if TRUE (FALSE by default) the total across 
all categories of \code{var} is included.
Only considered if \code{type} is 'countTable'.}

\item{rowVar}{Character vector with variable(s) used for the rows.
If multiple variables are specified, the variables should be sorted in hierarchical order.
The variables are included in rows, excepted if specified in \code{rowVarInSepCol}.}

\item{rowVarLab}{Label for the \code{rowVar} variable(s).}

\item{rowVarDataLevels}{Data.frame with levels to consider for the \code{rowVar} variable(s)
(in case of non nested row variable(s)).}

\item{rowOrder}{String or function of named list with method used to order the rows,
see the \code{method} parameter of \code{\link{convertVarToFactorWithOrder}}.
'total' is only available if the total is available for each \code{rowVar} (\code{rowSubtotalInclude} set to TRUE)
If a string, the same method is used for all \code{rowVar},
otherwise the list is named with the \code{rowVar} variable, to
specify a different ordering method for each variable.}

\item{rowOrderTotalFilterFct}{Function used to filter the data used to order the rows
based on total counts (in case \code{rowOrder} is 'total').}

\item{rowVarInSepCol}{Variable(s) of \code{rowVar} which should be 
included in separated column in the table.
NULL by default. 
This is only available if \code{rowVar} if not NULL.}

\item{rowTotalInclude}{Logical, if TRUE (FALSE by default) include the total
counts across rows in a separated row.}

\item{rowTotalLab}{label for the row with total}

\item{rowSubtotalInclude}{Logical, if TRUE (FALSE by default) include also the total for each
group of the nested variable specified in \code{rowVar}}

\item{rowInclude0}{Logical, if TRUE (FALSE by default),
include rows with no records, based on all combinations 
of the \code{rowVar} (assuming nested variable(s)).}

\item{emptyValue}{Value used to fill the table for missing values.
See the \code{fill} parameter of the \code{\link[reshape2]{dcast}} function.}

\item{colVar}{Character vector with variable(s) used for the columns.
If multiple variables are specified, the variables should be sorted in hierarchical order,
and are included in multi-columns layout.}

\item{colVarTotal}{String with column(s) considered to compute the total by,
reported in the header of the table, by default same as \code{colVar}.}

\item{colVarTotalPerc}{String with column(s) considered to compute the total by,
used as denominator for the percentage computation, by default same as \code{colVarTotal}.}

\item{colInclude0}{Logical, if TRUE (FALSE by default),
include columns with no records, based on all combinations 
of the \code{columnVar} (assuming nested variable(s)).
If variable(s) are not nested, possible combinations
can be specified via \code{colVarDataLevels}.}

\item{colVarDataLevels}{Data.frame with levels to consider for the \code{colVar} variable(s)
(in case of non nested column variable(s)).}

\item{colTotalInclude}{Logical, if TRUE (FALSE by default) include the summary 
statistics across columns in a separated column.}

\item{colTotalLab}{String, label for the total column included 
if \code{colTotalInclude} is TRUE, 'Total' by default.}

\item{subjectVar}{String, variable of \code{data} with subject ID,
'USUBJID' by default.}

\item{stats}{(optional) Named list of expression or call object of summary statistics of interest.
The following variables are recognized, if the table is a: 
\itemize{
\item{'summaryTable': }{
\itemize{
\item{'statN': }{number of subjects}
\item{'statMean': }{mean of \code{var}}
\item{'statSD': }{standard deviation of \code{var}}
\item{'statSE': }{standard error of \code{var}}
\item{'statMedian': }{median of \code{var}}
\item{'statMin': }{minimum of \code{var}}
\item{'statMax': }{maximum of \code{var}}
\item{'statPerc': }{percentage of subjects}
\item{'statm': }{number of records}
}
}
\item{'countTable': }{
\itemize{
\item{'statN': }{number of subjects}
\item{'statPercN': }{percentage of subjects}
\item{'statm': }{number of records}
}
}
}
If \code{stats} if of length 1, the name of the summary statistic is not included
in the table.
The statistics can be specified for each \code{var} (if multiple),
by specifying a nested list of such parameters named by the variable.}

\item{statsExtra}{(optional) Named list with functions for additional custom
statistics to be computed, only available for 'summaryTable',
e.g. list(statCVPercN = function(x) sd(x)/mean(x)*100).
Each function has as parameter: either 'x': the variable or 'data': the entire dataset,
and return the corresponding summary statistic.}

\item{statsValueLab}{String with label for the statistic value, 
in case no \code{colVar} is specified: 'StatisticValue' by default.}

\item{filterFct}{(optional) Function based on computed statistics of
\code{rowVar}/code{colVar} which returns a subset of the summary table of interest.
Note: the total count per category of the row variables is included
in a row with this variable set to 'Total'.}

\item{dataTotal}{Data.frame used to extract the Total count, indicated
in 'N' in column header.
It should contain the variables specified by \code{colVarTotal}.}

\item{dataTotalPerc}{Data.frame used to extract the Total count, 
used for the computation of the percentage: 'statPercN' parameter.
It should contain the variables specified by \code{colVarTotalPerc}.}

\item{type}{String with type of table: 
\itemize{
\item{'summaryTable': }{summary table with custom statistics}
\item{'countTable': }{count table}
\item{'auto' (by default): }{'summaryTable' of \code{var} is numeric,
'countTable' otherwise
}
}}

\item{labelVars}{Named string with variable labels (names are the variable code).}

\item{landscape}{Logical, if TRUE (by defaut) the table is presented in landscape
format.}

\item{margin}{Margin in the document in inches.}

\item{rowPadBase}{Base padding for row (number of spaces)}

\item{title}{Character vector with title(s) for the table.
Set to NULL (by default) if no title should be included.}

\item{footer}{Character vector with footer(s) for the table.
Set to NULL (by default) of no footer should be included.}

\item{file}{String with path of the file where the table should be exported.
If NULL, the summary table is not exported but only returned as output.}

\item{outputType}{String with output type, 'data.frame' or 'flextable'.}

\item{statsLayout}{String with layout for the statistics names 
(in case more than one statistic is included), among:
\itemize{
\item{row: }{Statistics are included in rows in the first column of the table}
\item{'col': }{Statistics are included in columns (last row of the header)}
\item{'rowInSepCol': }{Statistics are included in rows, but in a separated column than
the \code{rowVar} variable(s)}
}}

\item{style}{string with table style in case \code{outputType} is 'flextable',
either 'report' or 'presentation'}

\item{byVar}{Variable(s) of \code{data} for which separated table(s)
should be created.}

\item{byVarLab}{String with label for \code{byVar}, used to set the names
of the output list of table(s).}

\item{colHeaderTotalInclude}{Logical, if TRUE include the total of number of patients
(\code{'statN'}) in the header.}

\item{fontsize}{Integer with font size, 8 by default.}
}
\value{
Depending on the \code{outputType}:
\itemize{
\item{'flextable': }{\code{\link[flextable]{flextable}} object with summary table}
\item{'data.frame': }{data.frame with summary table}
}
If \code{summaryTable} is a list of summary tables,
returns a list of corresponding summary tables in long format.
}
\description{
Get summary statistics table
}
\author{
Laure Cougnaud
}
