% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryStatisticsTable.R
\name{getSummaryStatisticsTable}
\alias{getSummaryStatisticsTable}
\title{Get summary statistics table}
\usage{
getSummaryStatisticsTable(data, var = NULL, varLab = getLabelVar(var,
  labelVars = labelVars), varIgnore = NULL, rowVar = NULL,
  rowVarLab = getLabelVar(rowVar, labelVars = labelVars),
  rowVarInSepCol = NULL, rowTotalInclude = FALSE, rowTotalLab = NULL,
  colVar = NULL, subjectVar = "USUBJID", stats = NULL,
  dataTotal = NULL, type = "summaryTable", labelVars = NULL,
  landscape = FALSE, margin = 1, rowPadBase = 2,
  title = "Table: Descriptive statistics", footer = NULL,
  file = "summaryStatisticsTable.docx")
}
\arguments{
\item{data}{data.frame with data}

\item{var}{string, variable of \code{data} with variable to compute statistics on,
only used (and required) if \code{type} is 'summaryTable'.
Missing values, if present, are filtered.}

\item{varLab}{label for the \code{var} variable}

\item{varIgnore}{vector with elements to ignore in the \code{var} variable}

\item{rowVar}{character vector with variable(s) used for the rows.
If multiple variables are specified, the variables should be sorted in hierarchical order.
The variables are included in rows, excepted if specified in \code{rowVarInSepCol}.}

\item{rowVarLab}{label for the \code{rowVar} variable(s)}

\item{rowVarInSepCol}{variable(s) of \code{rowVar} which should be 
included in separated column in the table.
NULL by default. 
This is only available if \code{rowVar} if not NULL.}

\item{rowTotalInclude}{logical, if TRUE (FALSE by default) include the total
across rows in a separated row.}

\item{rowTotalLab}{label for the row with total}

\item{colVar}{character vector with variable(s) used for the columns.
If multiple variables are specified, the variables should be sorted in hierarchical order,
and are included in multi-columns layout.}

\item{subjectVar}{string, variable of \code{data} with subject ID,
'USUBJID' by default}

\item{stats}{(optional) named list of expression of summary statistics of interest.
The following statistics are recognized, if: 
\itemize{
\item{\code{type} is a 'summaryTable':}{'N', 'Mean', 'SD', 'SE', 'Median',
'Min', 'Max', 'Perc'}
\item{\code{type} is a 'countTable':}{'N','Perc'}
}
If \code{stats} if of length 1, the name of the summary statistic is not included
in the table.}

\item{dataTotal}{data.frame used to extract the Total count, indicated
in 'N' in column header and used for the computation of the percentage ('Perc') parameter.
It should contain the variables specified by \code{colVar}.}

\item{type}{string with type of summary table: 'summaryTable' 
(by default) or 'countTable'}

\item{labelVars}{named string with variable labels (names are the variable code)}

\item{landscape}{logical, if TRUE (by defaut) the table is presented in landscape
format}

\item{margin}{margin in the document in inches}

\item{rowPadBase}{base padding for row (number of spaces)}

\item{title}{character vector with title(s) for the table.
Set to NULL if no title should be included.}

\item{footer}{character vector with footer(s) for the table.
Set to NULL (by default) of no footer should be included.}

\item{file}{string with path of the file where the table should be exported.
If NULL, the summary table is not exported but only returned as output.}
}
\value{
\code{\link[flextable]{flextable}} object with summary table
}
\description{
Get summary statistics table
}
\author{
Laure Cougnaud
}
