% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryStatisticsTable.R
\name{getSummaryStatisticsTable}
\alias{getSummaryStatisticsTable}
\title{Get summary statistics table}
\usage{
getSummaryStatisticsTable(data, var = NULL, varLab = getLabelVar(var,
  data = data, labelVars = labelVars), varIgnore = NULL,
  varLabGeneral = "Variable", varLabSubgroup = "Subgroup",
  rowVar = NULL, rowVarLab = getLabelVar(rowVar, labelVars =
  labelVars), rowOrder = "auto", rowOrderTotalFilterFct = NULL,
  rowVarInSepCol = NULL, rowTotalInclude = FALSE, rowTotalLab = NULL,
  rowSubtotalInclude = FALSE, rowInclude0 = FALSE, colVar = NULL,
  colInclude0 = FALSE, colTotalInclude = FALSE,
  subjectVar = "USUBJID", stats = NULL, filterFct = NULL,
  dataTotal = NULL, type = "auto", labelVars = NULL,
  landscape = FALSE, margin = 1, rowPadBase = 2,
  title = "Table: Descriptive statistics", footer = NULL,
  file = "summaryStatisticsTable.docx", outputType = c("flextable",
  "data.frame"))
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{String, variable of \code{data} with variable to compute statistics on.
Missing values, if present, are filtered.}

\item{varLab}{Label for the \code{var} variable.}

\item{varIgnore}{Vector with elements to ignore in the \code{var} variable}

\item{varLabGeneral}{String with general label for variable specified in \code{var}.
In case of multiple variable in \code{var}, this will be included in the table header
(see 'rowVarLab' attribute of the output).}

\item{varLabSubgroup}{String with general label for sub-group, in case
\code{var} is specified for a count table.
This will be included in the table header (see 'rowVarLab' attribute of the output).}

\item{rowVar}{Character vector with variable(s) used for the rows.
If multiple variables are specified, the variables should be sorted in hierarchical order.
The variables are included in rows, excepted if specified in \code{rowVarInSepCol}.}

\item{rowVarLab}{Label for the \code{rowVar} variable(s).}

\item{rowOrder}{String or function of named list with method used to order the rows,
see the \code{method} parameter of \code{\link{convertVarToFactorWithOrder}}.
'total' is only available if the total is available for each \code{rowVar} (\code{rowSubtotalInclude} set to TRUE)
If a string, the same method is used for all \code{rowVar},
otherwise the list is named with the \code{rowVar} variable, to
specify a different ordering method for each variable.}

\item{rowOrderTotalFilterFct}{Function used to filter the data used to order the rows
based on total counts (in case \code{rowOrder} is 'total').}

\item{rowVarInSepCol}{Variable(s) of \code{rowVar} which should be 
included in separated column in the table.
NULL by default. 
This is only available if \code{rowVar} if not NULL.}

\item{rowTotalInclude}{Logical, if TRUE (FALSE by default) include the total
counts across rows in a separated row.}

\item{rowTotalLab}{label for the row with total}

\item{rowSubtotalInclude}{Logical, if TRUE (FALSE by default) include also the total for each
group of the nested variable specified in \code{rowVar}}

\item{rowInclude0}{Logical, if TRUE (FALSE by default),
include rows with \code{rowVar} elements not available in the data
(e.g. if a \code{rowVar} is a factor, include all levels even if no records in the data).}

\item{colVar}{Character vector with variable(s) used for the columns.
If multiple variables are specified, the variables should be sorted in hierarchical order,
and are included in multi-columns layout.}

\item{colInclude0}{Logical, if TRUE (FALSE by default),
include columns with \code{colVar} elements not available in the data
(e.g. if a \code{colVar} is a factor, include all levels even if no records in the data).}

\item{colTotalInclude}{Logical, if TRUE (FALSE by default) include the summary 
statistics across columns in a separated column.}

\item{subjectVar}{String, variable of \code{data} with subject ID,
'USUBJID' by default.}

\item{stats}{(optional) Named list of expression or call object of summary statistics of interest.
The following variables are recognized, if the table is a: 
\itemize{
\item{'summaryTable': }{
\itemize{
\item{'N': }{number of subjects}
\item{'Mean': }{mean of \code{var}}
\item{'SD': }{standard deviation of \code{var}}
\item{'SE': }{standard error of \code{var}}
\item{'Median': }{median of \code{var}}
\item{'Min': }{minimum of \code{var}}
\item{'Max': }{maximum of \code{var}}
\item{'Perc': }{percentage of subjects}
\item{'m': }{number of records}
\item{'Percm': }{percentage of records}
}
}
\item{'countTable': }{
\itemize{
\item{'N': }{number of subjects}
\item{'PercN': }{percentage of subjects}
\item{'m': }{number of records}
\item{'Percm': }{percentage of records}
}
}
}
If \code{stats} if of length 1, the name of the summary statistic is not included
in the table.}

\item{filterFct}{(optional) Function based on computed statistics of
\code{rowVar}/code{colVar} which returns a subset of the summary table of interest.
Note: the total count per category of the row variables is included
in a row with this variable set to 'Total'.}

\item{dataTotal}{Data.frame used to extract the Total count, indicated
in 'N' in column header and used for the computation of the percentage:
'PercN' and 'Percm' parameters.
It should contain the variables specified by \code{colVar}.}

\item{type}{String with type of table: 
\itemize{
\item{'summaryTable': }{summary table with custom statistics}
\item{'countTable': }{count table}
\item{'auto' (by default): }{'summaryTable' of \code{var} is numeric,
'countTable' otherwise
}
}}

\item{labelVars}{Named string with variable labels (names are the variable code).}

\item{landscape}{Logical, if TRUE (by defaut) the table is presented in landscape
format.}

\item{margin}{Margin in the document in inches.}

\item{rowPadBase}{Base padding for row (number of spaces)}

\item{title}{Character vector with title(s) for the table.
Set to NULL if no title should be included.}

\item{footer}{Character vector with footer(s) for the table.
Set to NULL (by default) of no footer should be included.}

\item{file}{String with path of the file where the table should be exported.
If NULL, the summary table is not exported but only returned as output.}

\item{outputType}{String with output type, 'data.frame' or 'flextable'.}
}
\value{
Depending on the \code{outputType}:
\itemize{
\item{'flextable': }{\code{\link[flextable]{flextable}} object with summary table}
\item{'data.frame': }{data.frame with summary table}
}
}
\description{
Get summary statistics table
}
\author{
Laure Cougnaud
}
