% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryStatisticsTable.R
\name{getSummaryStatisticsTable}
\alias{getSummaryStatisticsTable}
\title{Get summary statistics table}
\usage{
getSummaryStatisticsTable(data, var = NULL, varFlag = NULL,
  varLab = getLabelVar(var, data = data, labelVars = labelVars),
  varInclude0 = FALSE, varIgnore = NULL, varGeneralLab = "Variable",
  varSubgroupLab = NULL, varIncludeTotal = FALSE,
  varTotalInclude = FALSE, varTotalInSepRow = FALSE, rowVar = NULL,
  rowVarLab = getLabelVar(rowVar, labelVars = labelVars),
  rowVarDataLevels = NULL, rowOrder = "auto",
  rowOrderTotalFilterFct = NULL, rowOrderCatLast = "Other",
  rowVarFormat = NULL, rowVarTotalInclude = NULL,
  rowVarTotalByVar = NULL, rowVarTotalInSepRow = NULL,
  rowTotalLab = NULL, rowInclude0 = FALSE, rowAutoMerge = TRUE,
  emptyValue = "-", rowVarTotalPerc = NULL, colVar = NULL,
  colVarTotal = colVar, colVarTotalPerc = colVarTotal,
  colInclude0 = FALSE, colVarDataLevels = NULL,
  colTotalInclude = FALSE, colTotalLab = "Total", stats = NULL,
  statsExtra = NULL, statsVarBy = NULL,
  statsGeneralLab = "Statistic", statsValueLab = "StatisticValue",
  subjectVar = "USUBJID", filterFct = NULL, dataTotal = NULL,
  dataTotalPerc = dataTotal, dataTotalRow = NULL,
  dataTotalCol = NULL, type = "auto", labelVars = NULL,
  landscape = (style == "presentation"), margin = 1,
  rowPadBase = 14.4, title = NULL, footer = NULL, file = NULL,
  outputType = "flextable", statsLayout = ifelse(outputType ==
  "flextable", "row", "col"), style = "report",
  colorTable = getColorTable(style = style), byVar = NULL,
  byVarLab = getLabelVar(byVar, data = data, labelVars = labelVars),
  colHeaderTotalInclude = TRUE, fontsize = switch(style, report = 8,
  presentation = 10), fontname = switch(style, report = "Times",
  presentation = "Tahoma"), vline = "none", pageDim = NULL,
  expandVar = NULL, noEscapeVar = NULL, barVar = NULL, ...)
}
\arguments{
\item{data}{Data.frame with data.}

\item{var}{Character vector, variable(s) of \code{data} to compute statistics on.
Missing values, if present, are filtered.}

\item{varFlag}{Character vector, subset of \code{var} with variable(s) of type 'flag' (with 'Y' or 'N').
Only the counts for records flagged (with 'Y') are retained.}

\item{varLab}{Label for the \code{var} variable.}

\item{varInclude0}{Logical, if TRUE (FALSE by default)
include rows with no counts for the count \code{var} or \code{varFlag} variable(s).}

\item{varIgnore}{Vector with elements to ignore in the \code{var} variable}

\item{varGeneralLab}{String with general label for variable specified in \code{var}.
In case of multiple variable in \code{var}, this will be included in the table header
(see 'rowVarLab' attribute of the output).}

\item{varSubgroupLab}{String with general label for sub-group, in case
\code{var} is specified for a count table.
This will be included in the table header (see 'rowVarLab' attribute of the output).
Empty by default.}

\item{varIncludeTotal}{This argument is deprecated, please use: 'varTotalInclude' instead.}

\item{varTotalInclude}{Should the total across all categories of \code{var} 
be included for the count table?
Either:
\itemize{
\item{logical of length 1, if TRUE (FALSE by default) include the total for all \code{var}}
\item{a character vector containing \code{var} for which the total should be included}
}}

\item{varTotalInSepRow}{Logical, should the total per variable be included in
a separated row (by default) or in the row containing the header of the variable?}

\item{rowVar}{Variable(s) of \code{data} used for
grouping in row in the final table.}

\item{rowVarLab}{Label for the \code{rowVar} variable(s).}

\item{rowVarDataLevels}{Data.frame with unique combinations of \code{rowVar}
to be included in columns.
Each column should correspond to \code{colVar} and as factor
if the elements should be ordered in the final table.}

\item{rowOrder}{Specify how the rows should be ordered in the table, either a:
\itemize{
\item{String among:}{
\itemize{
\item{'auto' (by default): }{if the variable is a factor, keep its order, otherwise order alphabetically}
\item{'alphabetical': }{order alphabetically}
\item{'total': }{order rows in decreasing order of the total number of subjects
across all columns for this specific category.}
}}
\item{Function with input the summary table and output the ordered elements of the \code{rowVar}}
}
To specify different ordering methods for different \code{rowVar}, specify a list
of such elements, named with the \code{rowVar} variable.}

\item{rowOrderTotalFilterFct}{Function used to filter the data used to order the rows
based on total counts (in case \code{rowOrder} is 'total'),
To order rows based on one specific column category,
e.g. to order based on the counts in the treatment column:
function(x) subset(x, TRTP == "treatmentX")}

\item{rowOrderCatLast}{String with category to be printed last of each \code{rowVar}
(if any, set to NULL if none). By default, category labelled 'Other' is printed last.}

\item{rowVarFormat}{Named list with special formatting for the \code{rowVar}.
Currently, only possibility is to set the variable elements in bold, with:
list(var1 = "bold").
(Use 'variable' for \code{var} or 'variableGroup' for group within categorical variables.)}

\item{rowVarTotalInclude}{Character vector with \code{rowVar}
for which to include the total for each group.}

\item{rowVarTotalByVar}{Character vector with extra \code{rowVar}
for which to compute the row total by.
Can be specified for each \code{rowVarTotalInclude} if named by the corresponding variable.}

\item{rowVarTotalInSepRow}{Character vector with \code{rowVarTotalInclude}
(not in \code{rowVarInSepCol}) for which the total should be included in a separated row labelled 'Total'.
Otherwise (by default) the total is included in the header row of each category.}

\item{rowTotalLab}{label for the row with total}

\item{rowInclude0}{Logical, if TRUE (FALSE by default),
include rows with no records, based on all combinations 
of the \code{rowVar} (assuming nested variable(s)).}

\item{rowAutoMerge}{Logical, if TRUE (by default) automatically merging of rows,
e.g. in case there is only one sub-category (e.g. categorical variable with only one group)
or only one statistic per category}

\item{emptyValue}{Value used to fill the table for missing values, '-' by default.
See the \code{fill} parameter of the \code{\link[reshape2]{dcast}} function.}

\item{rowVarTotalPerc}{Character vector with row(s) considered to compute the total,
used as denominator for the percentage computation.
By default the total is only computed by column (NULL by default).
If the total should be based on the total number of records per variable,
\code{rowVarTotalPerc} should be set to 'variable'.}

\item{colVar}{Character vector with variable(s) used for the columns.
If multiple variables are specified, the variables should be sorted in hierarchical order,
and are included in multi-columns layout.}

\item{colVarTotal}{String with column(s) considered to compute the total by,
reported in the header of the table, by default same as \code{colVar}.}

\item{colVarTotalPerc}{String with column(s) considered to compute the total by,
used as denominator for the percentage computation, by default same as \code{colVarTotal}.}

\item{colInclude0}{Logical, if TRUE (FALSE by default),
include columns with no records, based on all combinations 
of the \code{columnVar} (assuming nested variable(s)).
If variable(s) are not nested, possible combinations
can be specified via \code{colVarDataLevels}.}

\item{colVarDataLevels}{Data.frame with unique combinations of \code{colVar}
to be included in columns.
Each column should correspond to \code{colVar} and as factor
if the elements should be ordered in the final table.}

\item{colTotalInclude}{Logical, if TRUE (FALSE by default) include the summary 
statistics across columns in a separated column.}

\item{colTotalLab}{String, label for the total column included 
if \code{colTotalInclude} is TRUE, 'Total' by default.}

\item{stats}{(Optionally) Either:
\itemize{
\item{string with: }{
\itemize{
\item{'default': }{default sets of statistics, 
see types: 'summary-default' and 'count-default' in \code{\link{getStats}}}
\item{'all': }{all computed statistics, see types: 'summary' and 'count' in \code{\link{getStats}}}
}}
\item{Named list of expressions or call objects of summary statistics of interest: }{
The names are reported in the header.
The following variables are recognized, if the table is a: 
\itemize{
\item{'summaryTable': }{
\itemize{
\item{'statN': }{number of subjects}
\item{'statMean': }{mean of \code{var}}
\item{'statSD': }{standard deviation of \code{var}}
\item{'statSE': }{standard error of \code{var}}
\item{'statMedian': }{median of \code{var}}
\item{'statMin': }{minimum of \code{var}}
\item{'statMax': }{maximum of \code{var}}
\item{'statPerc': }{percentage of subjects}
\item{'statm': }{number of records}
}
}
\item{'countTable': }{
\itemize{
\item{'statN': }{number of subjects}
\item{'statPercN': }{percentage of subjects}
\item{'statm': }{number of records}
}
}
}
If \code{stats} if of length 1, the name of the summary statistic is not included
in the table.
The statistics can be specified for each \code{var} (if multiple), 
by naming each element of the list:
list(varName1 = list(...), varName2 = list()) and/or for each element in:
\code{statsVarBy}, by naming each sublist.
}}}

\item{statsExtra}{(optional) Named list with functions for additional custom
statistics to be computed, only available for 'summaryTable',
e.g. list(statCVPerc = function(x) sd(x)/mean(x)*100) (or \code{\link{cv}}).
Each function has as parameter: either 'x': the variable or 'data': the entire dataset,
and return the corresponding summary statistic.}

\item{statsVarBy}{String with variable in \code{rowVar}/code{colVar}
which the statistics should be computed by.
In this case, \code{stats} (nested list or not) should be additionally nested
to specify the statistics for each element in \code{statsVarBy}.}

\item{statsGeneralLab}{String with general label for statistics, 'Statistic' by default.
Only included if no \code{statsVar} if longer than 1.}

\item{statsValueLab}{String with label for the statistic value, 
'StatisticValue' by default.
This is only used if the statistics provided in \code{stats} are not named
and if no \code{colVar} is specified.}

\item{subjectVar}{String, variable of \code{data} with subject ID,
'USUBJID' by default.}

\item{filterFct}{(optional) Function based on computed statistics of
\code{rowVar}/code{colVar} which returns a subset of the summary table 
(after statistics computation).
Note: The filtering function should also handle records with :
\itemize{
\item{\code{rowVar}/\code{codeVar} set to 'Total'/\code{colTotalLab} 
if \code{rowVarTotalInclude}/\code{colTotalInclude} is TRUE}
\item{total for the column header: \code{isTotal} set to TRUE,
and \code{colVar}/\code{rowVar} is NA}
}
Note: the total count per category of the row variables is included
in a row with this variable set to 'Total'.}

\item{dataTotal}{Data.frame used to extract the Total number of subject
per column in column header ('N = [X]').
It should contain the variables specified by \code{colVarTotal}.
If not specified, the total number of subjects is extracted from the \code{data}.}

\item{dataTotalPerc}{Data.frame used to extract the Total count per column 
for the computation of the percentage ('statPercN' statistic).
By default same as \code{dataTotal} .
It should contain the variables specified by \code{colVarTotalPerc}.}

\item{dataTotalRow}{Data.frame used to extract the total count across all
elements of the row
variable, in case \code{rowVarTotalInclude} is specified,
or list of such data.frame for each \code{rowVar} variable (named by variable).}

\item{dataTotalCol}{Data.frame from which the total across columns is 
extracted (in case \code{colTotalInclude} is TRUE)
or list of such data.frame for each \code{rowVar} variable.
This data is used for:
\itemize{
\item{the header of the total column in case \code{dataTotal} is
not specified}
\item{the denominator of the percentages in the total column
in case \code{dataTotalPerc} is not specified}
}}

\item{type}{String with type of table: 
\itemize{
\item{'summaryTable': }{summary table with custom statistics}
\item{'countTable': }{count table}
\item{'auto' (by default): }{'summaryTable' of \code{var} is numeric,
'countTable' otherwise
}
}}

\item{landscape}{Logical, if TRUE (by defaut) the table is presented in landscape
format.}

\item{margin}{Margin in the document in inches, currently only used to specify the
width of the table: [width page extracted from \code{dimPage} - 2* margin]}

\item{rowPadBase}{Base padding for row (in points), 14.4 by default (corresponds to 0.2 inches)}

\item{title}{Character vector with title(s) for the table.
Set to NULL (by default) if no title should be included.
If vector > 1, specified for each element of \code{byVar} (in order of the levels).}

\item{footer}{Character vector with footer(s) for the table.
Set to NULL (by default) of no footer should be included.}

\item{file}{String with path of the file where the table should be exported.
If NULL, the summary table is not exported but only returned as output.}

\item{outputType}{String with output type, 'data.frame' or 'flextable'.}

\item{statsLayout}{String with layout for the statistics names 
(in case more than one statistic is included), among:
\itemize{
\item{row (by default when \code{outputType} is: 'flextable'): }{
All statistics are included in different rows 
in the first column of the table
}
\item{'col' (by default when \code{outputType} is: 'DT'): }{
Statistics are included in separated columns (last row of the header).
This option is not compatible with categorical variable(s).
}
\item{'rowInSepCol': }{
Statistics are included in different rows, but in a separated column than
the \code{rowVar} variable(s)
}
}}

\item{style}{string with table style in case \code{outputType} is 'flextable',
 either 'report' or 'presentation'.
This parameter affects the fontsize, font family, color of the text and background, 
and table dimensions of the table.}

\item{colorTable}{Named character vector with color for the table background/body/text/line,
e.g. created with the \code{\link[glpgStyle]{getColorTable}} function.}

\item{byVar}{Variable(s) of \code{data} for which separated table(s)
should be created.}

\item{byVarLab}{String with label for \code{byVar}, used to set the names
of the output list of table(s).}

\item{colHeaderTotalInclude}{Logical, if TRUE include the total of number of patients
(\code{'statN'}) in the header.}

\item{fontsize}{Integer with font size, by default:
8 if \code{style} is 'report' and 10 if \code{style} is 'presentation'.}

\item{fontname}{String with font name, by default:
'Times' if \code{style} is 'report' and 'Tahoma' if \code{style} is 'presentation'.}

\item{vline}{String mentioning how vertical lines should be included, either: 
\itemize{
\item{'none' (default): }{no vertical lines included}
\item{'auto': }{vertical lines included between sub-groups}
}}

\item{pageDim}{Numeric vector of length 2 with page width and height
depending on \code{outputType}:
\itemize{
\item{'flextable': }{in inches in portrait format}
\item{'DT': }{number of rows in the table (currently only
the height is used (e.g. \code{c(NA, 4)})
}}}

\item{expandVar}{Character vector with variables of the summary table which
should be expanded in the data (only for 'DT' output).}

\item{noEscapeVar}{Character vector with variables of \code{summaryTable}
which shouldn't be escaped in the table (e.g. containing URLs).}

\item{barVar}{Character vector with variables of \code{summaryTable}
that should be represented as a bar.}

\item{...}{Extra parameters passed to the 
\code{\link[glpgUtilityFct]{toDTGLPG}} (only for 'DT' output)}
}
\value{
Depending on the \code{outputType}:
\itemize{
\item{'data.frame-base': }{input summary table in a long format with
all computed statistics}
\item{'data.frame': }{summary table in a wide format (
different columns for each \code{colVar}), with specified labels}
\item{'flextable' (by default): }{\code{\link[flextable]{flextable}}
 object with summary table}
\item{'DT': }{\code{\link[DT]{datatable}} object with summary table}
}
If multiple \code{outputType} are specified, a list of those objects, named
by \code{outputType}.
If \code{byVar} is specified, each object consists of a list of tables,
one for each element in \code{byVar}.
}
\description{
Get summary statistics table
}
\author{
Laure Cougnaud
}
