% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-utility.R
\name{inTextSummaryTable-stats-utility}
\alias{inTextSummaryTable-stats-utility}
\title{Common arguments for the for the statistics utility functions
of the inTextSummaryTable package.}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Logical, should NA value(s) be removed (FALSE by default)?}
}
\value{
No return value, used for the documentation of 
stat utility R functions
}
\description{
Common arguments for the for the statistics utility functions
of the inTextSummaryTable package.
}
