% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{roundCustom}
\alias{roundCustom}
\title{Custom round function, rounding to the closest digits
(instead of rounding to the even number in case of 0.5)}
\usage{
roundCustom(x, digits)
}
\arguments{
\item{x}{numeric vector to round}

\item{digits}{number of digits to consider}
}
\value{
rounded vector
}
\description{
Custom round function, rounding to the closest digits
(instead of rounding to the even number in case of 0.5)
}
\author{
stackoverflow question 6461209
}
