% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{roundCustomText}
\alias{roundCustomText}
\title{Custom round function, rounding to the closest digits
(instead of rounding to the even number in case of 0.5)}
\usage{
roundCustomText(..., format = c("text", "number"))
}
\arguments{
\item{...}{Additional parameters for the 
\code{\link[glpgUtilityFct]{roundCustom}} function.}

\item{format}{string with format for the number: 
'text' (with trailing zeros) (by default) or 'number'.}
}
\description{
Custom round function, rounding to the closest digits
(instead of rounding to the even number in case of 0.5)
}
