% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectProfileSummaryPlot.R
\name{subjectProfileSummaryPlot}
\alias{subjectProfileSummaryPlot}
\title{Plot subject summary profile}
\usage{
subjectProfileSummaryPlot(data, xVar = NULL, xLab = getLabelVar(xVar,
  labelVars = labelVars), meanVar = "statMean", seVar = if ("statSE"
  \%in\% colnames(data)) "statSE", yLab = paste(sub("^stat", "",
  meanVar), if (!is.null(seVar)) paste("+-", sub("^stat", "", seVar))),
  facetVar = NULL, facetScale = "free_y", colorVar = NULL,
  colorLab = getLabelVar(colorVar, labelVars = labelVars),
  colorPalette = NULL, labelVars = NULL, useLinetype = TRUE,
  linetypePalette = NULL, useShape = TRUE, shapePalette = NULL,
  jitter = NULL, title = NULL, yLim = NULL, xLim = NULL,
  xAxisLabs = NULL, sizePoint = GeomPoint$default_aes$size,
  sizeLine = GeomLine$default_aes$size,
  sizeLabel = GeomText$default_aes$size,
  widthErrorBar = GeomErrorbar$default_aes$width, tableText = NULL,
  tableLabel = NULL, tableHeight = 0.1, label = FALSE,
  labelPadding = unit(1, "lines"), byVar = NULL, hLine = NULL,
  hLineColor = "black", hLineLty = "solid", vLine = NULL,
  vLineColor = "black", vLineLty = "solid", style = "report",
  fontname = switch(style, report = "Times", presentation = "Tahoma"),
  fontsize = switch(style, report = 8, presentation = 10),
  themeFct = switch(style, report = theme_classic, presentation =
  theme_bw), themeIncludeVerticalGrid = TRUE, ggExtra = NULL)
}
\arguments{
\item{data}{Data.frame with summary statistics to represent in the plot,
e.g. produced with the \code{\link{computeSummaryStatisticsTable}}.}

\item{xVar}{String, variable of \code{data} with variable for the x-axis.}

\item{xLab}{String with label for the \code{xVar}.}

\item{meanVar}{String, variable of \code{data} with the mean variable.}

\item{seVar}{String, variable of \code{data} with the standard error.}

\item{yLab}{String with label for the y-axis.
If different labels should be used for different elements of
\code{byVar} variable, the vector should be named
with each corresponding element (collapsed with '.' if multiple).}

\item{facetVar}{String, variable of \code{data} for facetting.}

\item{facetScale}{String with type of scale used for facetting, 'free_y' by default
(fixed scale in the x-axis and free in the y-axis).}

\item{colorVar}{String, variable of \code{data} for coloring.}

\item{colorLab}{String, label for \code{colorVar}, used in the legend.}

\item{colorPalette}{(named) Vector with color palette.}

\item{labelVars}{Named string with variable labels (names are the variable code).}

\item{useLinetype}{Logical, if TRUE (FALSE by default) use also linetype
to differenciate the variable specified via \code{colorVar} in the mean line.}

\item{linetypePalette}{Vector with linetype(s), in case \code{useLinetype} is TRUE.}

\item{useShape}{Logical, if TRUE (by default) \code{colorVar} is also used for the shape.}

\item{shapePalette}{Named vector with shape palette for \code{colorVar}.}

\item{jitter}{Numeric with jitter for the x-axis, only used if \code{colorVar} specified.}

\item{title}{String with title for the plot.
If different labels should be used for different elements of
\code{byVar} variable, the vector should be named
with each corresponding element (collapsed with '.' if multiple).}

\item{yLim}{Vector of the length 2 with limits for the y-axis.}

\item{xLim}{Vector of the length 2 with limits for the x-axis.}

\item{xAxisLabs}{(optional) Named character vector with labels for the x-axis.}

\item{sizePoint}{Size for the point.}

\item{sizeLine}{Size for the line linking means and error bars.}

\item{sizeLabel}{Size for the label, only used if \code{label} is not NULL.}

\item{widthErrorBar}{Numeric vector of length 1 with width of error bar.}

\item{tableText}{(optional) Character vector with colname of \code{data}
or expression from colnames of \code{data} to be represented in
the table below the plot.
By default, no table is displayed.}

\item{tableLabel}{Character vector with label for \code{tableText},
used as label for the x-axis of the table.}

\item{tableHeight}{Numeric of length 1 with height for the table.}

\item{label}{Logical or expression.
Points are labelled with \code{meanVar} if set to TRUE,
or with the specified expression if \code{label} is an expression.}

\item{labelPadding}{Amount of padding around labelled points,
1.5 lines by default, see parameter \code{point.padding} of the
\code{\link[ggrepel]{geom_text_repel}} function.}

\item{byVar}{Variable(s) of \code{data} for which separated plot(s)
should be created.}

\item{hLine}{(optional) numeric with y-intercept of dashed line to be added.
If different thresholds should be used for different elements of the 
\code{byVar} or \code{facetVar} variables, the vector should be named
with each corresponding element (collapsed with '.' if multiple).}

\item{hLineColor}{String with color for \code{hLine}.}

\item{hLineLty}{String with linetype for \code{hLine}.}

\item{vLine}{(optional) numeric with x-intercept of dashed line to be added.
If different thresholds should be used for different elements of the 
\code{byVar} or \code{facetVar} variables, the vector should be named
with each corresponding element (collapsed with '.' if multiple).}

\item{vLineColor}{String with color for \code{vLine}.}

\item{vLineLty}{String with linetype for \code{vLine}.}

\item{style}{String with subject profile style, either 'report' or 'presentation'.}

\item{fontname}{String with font name.}

\item{fontsize}{Numeric vector with font size.}

\item{themeFct}{Function with ggplot2 theme.}

\item{themeIncludeVerticalGrid}{Logical, if TRUE (by default)
include theme vertical grid lines (if present in \code{themeFct}).}

\item{ggExtra}{Extra \code{ggplot} command to be added in main plot.}
}
\value{
\code{\link[ggplot2]{ggplot}} object or list of such
objects of \code{byVar} is specified.
}
\description{
Plot subject summary profile
}
\author{
Laure Cougnaud
}
