% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableColors.R
\docType{data}
\name{tableColorsPresentation}
\alias{tableColorsPresentation}
\title{Colors for tables in a presentation style}
\format{
An object of class \code{character} of length 11.
}
\usage{
tableColorsPresentation
}
\description{
Default colors are
\itemize{
\item{header: }{white text on a blue background}
\item{body: }{black text on a grey background}
\item{footer: }{black text on a white background.}
}
}
\keyword{datasets}
